Commons

The commons section of the documentation collects common elements used by more than one solver.

Overview

debug actions that can be used in any solver, including the DebugActionFactory that turns the input array to a debug= keyword of a solver into the corresponding concrete debugs, especially when passing symbols like :Cost.

stopping criteria that can be used with different solvers in the stopping_criterion= keyword.

Passing parameters

Since the overall design of Manopt.jl is modular, one way to set parameters, for example in the objective of a sub problem, is via set_parameter!.

Manopt.is_tutorial_mode — Method

is_tutorial_mode()

A small internal helper to indicate whether tutorial mode is active.

You can set the mode by calling set_parameter!(:Mode, "Tutorial") or deactivate it by set_parameter!(:Mode, "").