Extension with JuMP.jl

Manopt can be used using the JuMP.jl interface, see the tutorial TODO:

Setting s solver and its options

A main thing to choose is the solver to use. By default this is set to the GradientDescentState. To change the solver you can set it with set_attribute. For exmple to use the quasi_Newton instead, use

model =  Model(Manopt.JuMP_Optimizer)
set_attribute(model, "descent_state_type", Manopt.QuasiNewtonState)

Any of the keywords of the solver you can set with set_attribute(model, keyword, value) for example to change the retraction to use, call

set_attribute(model, "retraction_method", ManifoldsBase.ProjectionRetraction())

Interface functions

Several functions from the Mathematical Optimization Interface (MOI) are extended when both `Manopt.jl and JuMP.jl are loaded:

Manopt.JuMP_Optimizer — Function

Manopt.JuMP_Optimizer()

Represent a solver from Manopt.jl within the MathOptInterface (MOI) framework.

See ManoptOptimizer for the fields and their meaning.

Manopt.JuMP_ManifoldSet — Function

JuMP_ManifoldSet(M::ManifoldsBase.AbstractManifold) = ManifoldSet(M)

Create a ManifoldSet.

Internal functions

JuMP.build_variable — Function

JuMP.build_variable(::Function, func, manifold::ManifoldsBase.AbstractManifold)

Build a JuMP.VariablesConstrainedOnCreation object containing variables and the ManifoldSet in which they should belong as well as the shape that can be used to go from the vectorized MOI representation to the shape of the manifold, that is, a ManifoldArrayShape.

JuMP.jump_function — Function

JuMP.jump_function(::JuMP.AbstractModel, F::Type{<:RiemannianFunction})

The JuMP function of a RiemannianFunction for any AbstractModel is that function itself.

JuMP.jump_function_type — Function

JuMP.jump_function_type(::JuMP.AbstractModel, F::Type{<:RiemannianFunction})

The JuMP function type of a function of type RiemannianFunction for any AbstractModel is that function type itself

JuMP.set_objective_function — Function

JuMP.set_objective_function(model::JuMP.Model, obj::Manopt.AbstractManifoldObjective)

Set the objective function of a JuMP.Model model to an AbstractManifoldObjective obj. This allows to use @objective with an objective from Manopt.jl.

MathOptInterface.add_constrained_variables — Function

MOI.add_constrained_variables(model::ManoptOptimizer, set::ManifoldSet)

Add MOI.dimension(set) variables constrained in set and return the list of variable indices that can be used to reference them as well a constraint index for the constraint enforcing the membership of the variables in the ManifoldSet set.

MathOptInterface.copy_to — Function

MOI.copy_to(dest::ManoptOptimizer, src::MOI.ModelLike)

Because supports_incremental_interface(dest) is true, this simply uses MOI.Utilities.default_copy_to and copies the variables with MOI.add_constrained_variables and the objective sense with MOI.set.

MathOptInterface.empty! — Function

MOI.empty!(model::ManoptJuMPExt.Optimizer)

Clear all model data from model but keep the options set.

MathOptInterface.dimension — Function

MOI.dimension(set::ManifoldSet)

Return the representation side of points on the (vectorized in representation) manifold. As the MOI variables are real, this means if the representation_size yields (in product) n, this refers to the vectorized point / tangent vector from (a subset of $ℝ^n$).

MathOptInterface.Utilities.map_indices — Function

MOI.Utilities.map_indices(index_map::Function, func::RiemannianFunction)

The original docstring states something about substituting some variable indices by their index map variants. On a RiemannianFunction there is nothing to substitute,

MathOptInterface.supports_add_constrained_variables — Function

MOI.supports_add_constrained_variables(::JuMP_Optimizer, ::Type{<:ManifoldSet})

Return true indicating that Manopt.JuMP_Optimizer support optimization on variables constrained to belong in a vectorized manifold ManifoldSet.

MathOptInterface.get — Function

MOI.get(::ManoptOptimizer, ::MOI.SolverVersion)

Return the version of the Manopt solver, it corresponds to the version of Manopt.jl.

MOI.get(model::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)

Return last value set by MOI.set(model, attr, value).

MOI.get(::ManoptOptimizer, ::MOI.SolverName)

Return the name of the Optimizer with the value of the descent_state_type option.

MOI.get(model::ManoptOptimizer, ::MOI.NumberOfVariables)

Return the number of variables added in the model, this corresponds to the MOI.dimension of the ManifoldSet.

MOI.get(model::ManoptOptimizer, ::MOI.ObjectiveSense)

Return the objective sense, defaults to MOI.FEASIBILITY_SENSE if no sense has already been set.

MOI.get(model::ManoptOptimizer, ::MOI.ResultCount)

Return MOI.OPTIMIZE_NOT_CALLED if optimize! hasn't been called yet and MOI.LOCALLY_SOLVED otherwise indicating that the solver has solved the problem to local optimality the value of MOI.RawStatusString for more details on why the solver stopped.

MOI.get(model::ManoptOptimizer, ::MOI.ResultCount)

Return 0 if optimize! hasn't been called yet and 1 otherwise indicating that one solution is available.

MOI.get(model::ManoptOptimizer, ::MOI.PrimalStatus)

Return MOI.NO_SOLUTION if optimize! hasn't been called yet and MOI.FEASIBLE_POINT otherwise indicating that a solution is available to query with MOI.VariablePrimalStart.

MOI.get(::ManoptOptimizer, ::MOI.DualStatus)

Returns MOI.NO_SOLUTION indicating that there is no dual solution available.

MOI.get(model::ManoptOptimizer, ::MOI.RawStatusString)

Return a String containing Manopt.get_reason without the ending newline character.

MOI.get(model::ManoptOptimizer, attr::MOI.ObjectiveValue)

Return the value of the objective function evaluated at the solution.

MOI.get(model::ManoptOptimizer, attr::MOI.VariablePrimal, vi::MOI.VariableIndex)

Return the value of the solution for the variable of index vi.

MathOptInterface.is_valid — Function

MOI.is_valid(model::ManoptOptimizer, vi::MOI.VariableIndex)

Return whether vi is a valid variable index.

MathOptInterface.supports — Function

MOI.supports(::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)

Return a Bool indicating whether attr.name is a valid option name for Manopt.

MOI.supports(::Manopt.JuMP_Optimizer, attr::MOI.RawOptimizerAttribute)

Return true indicating that Manopt.JuMP_Optimizer supports starting values for the variables.

MOI.supports(::ManoptOptimizer, ::Union{MOI.ObjectiveSense,MOI.ObjectiveFunction})

Return true indicating that Optimizer supports being set the objective sense (that is, min, max or feasibility) and the objective function.

MathOptInterface.supports_incremental_interface — Function

MOI.supports_incremental_interface(::JuMP_Optimizer)

Return true indicating that Manopt.JuMP_Optimizer implements MOI.add_constrained_variables and MOI.set for MOI.ObjectiveFunction so it can be used with JuMP.direct_model and does not require a MOI.Utilities.CachingOptimizer. See MOI.supports_incremental_interface.

MathOptInterface.set — Function

MOI.get(model::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)

Set the value for the keyword argument attr.name to give for the constructor model.options[DESCENT_STATE_TYPE].

function MOI.set(
    model::ManoptOptimizer,
    ::MOI.VariablePrimalStart,
    vi::MOI.VariableIndex,
    value::Union{Real,Nothing},
)

Set the starting value of the variable of index vi to value. Note that if value is nothing then it essentially unset any previous starting values set and hence MOI.optimize! unless another starting value is set.

MOI.set(model::ManoptOptimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSense)

Modify the objective sense to either MOI.MAX_SENSE, MOI.MIN_SENSE or MOI.FEASIBILITY_SENSE.

MOI.set(model::ManoptOptimizer, ::MOI.ObjectiveFunction{F}, func::F) where {F}

Set the objective function as func for model.

Internal wrappers and their functions

ManoptJuMPExt.ManifoldArrayShape — Type

ManifoldArrayShape{N} <: JuMP.AbstractShape

Represent some generic AbstractArray of a certain size representing an element on a manifold, that is either a point or a tangent vector.

Fields

size::NTuple{N,Int}: The size of the array

ManoptJuMPExt.ManifoldPointShape — Type

ManifoldPointShape{M<:ManifoldsBase.AbstractManifold,P}

Given a concrete manifold and a point type AbstractManifoldPoint, this type can represent all information necessary to “reshape” or “vectorize” such a point or to transform a vector back into a point of type P on M.

Fields

manifold::M: The manifold on which the point resides

Constructor

ManifoldPointShape(M::TM, ::Type{P<:ManifoldsBase.AbstractManifoldPoint})

Create a shape of a point on the manifold M of type P

ManoptJuMPExt.ManifoldSet — Type

ManifoldSet{M<:ManifoldsBase.AbstractManifold} <: MOI.AbstractVectorSet

Model a manifold from ManifoldsBase.jl as a vectorial set in the MathOptInterface (MOI). This is a slight misuse of notation, since the manifold itself might not be embedded, but just be paremetrized in a certain way.

Fields

manifold::M: The manifold in which the variables are constrained to lie. This is a ManifoldsBase.AbstractManifold object.

ManoptJuMPExt.ManoptOptimizer — Type

ManoptOptimizer <: MOI.AbstractOptimizer

Represent a solver from Manopt.jl within the MathOptInterface (MOI) framework.

Fields

problem::([AbstractManoptProblem`](@ref) a problem in manopt, especially containing the manifold and the objective function. It can be constructed as soon as the manifold and the objective are present.

manifold::(AbstractManifold the manifold on which the optimization is performed.

objective::(AbstractManifoldObjective the objective function to be optimized.

state::(AbstractManoptSolverState the state specifying the solver to use.

variable_primal_start::Vector{Union{Nothing,Float64}} starting value for the solver, in a vectorized form that JuMP.jl requires.

sense::MOI.OptimizationSense the sense of optimization, currently only minimization and maximization are supported.

options::Dict{String,Any}: parameters specifying a solver before the state is initialized, so especially which AbstractManoptSolverState to use, when setting up the `state.

All types in brackets can also be Nothing, indicating they were not yet initialized.

TODO: We might have to store the point and vector types P and T here maybe?

at least for the case where we are not on ManifoldArrayShape

ManoptJuMPExt.RiemannianFunction — Type

RiemannianFunction{MO<:Manopt.AbstractManifoldObjective} <: MOI.AbstractScalarFunction

A wrapper for a AbstractManifoldObjective that can be used as a MOI.AbstractScalarFunction.

Fields

func::MO: The AbstractManifoldObjective function to be wrapped.

ManoptJuMPExt.TangentVectorShape — Type

TangentVectorShape{M<:ManifoldsBase.AbstractManifold,T}

Represent a tangect vector on the tangent bundle of a manifold M.

Fields

manifold::M: The manifold on which the tangent vector resides

p::P (optioal) the base point P the tangent space is at where the vector of type T is defined.

Constructor

TangentVectorShape(M::TM, ::Type{T<:ManifoldsBase.AbstractTangentVector}, p::P=nothing)

Create a shape of a tangent vector on the manifold M of type T, where optionally a base point p can be specified.

Base.length — Method

length(shape::ManifoldArrayShape)

Return the length of the vectors in the vectorized representation.

JuMP.build_variable — Method

JuMP.build_variable(::Function, func, manifold::ManifoldsBase.AbstractManifold)

Build a JuMP.VariablesConstrainedOnCreation object containing variables and the ManifoldSet in which they should belong as well as the shape that can be used to go from the vectorized MOI representation to the shape of the manifold, that is, a ManifoldArrayShape.

JuMP.jump_function — Method

JuMP.jump_function(::JuMP.AbstractModel, F::Type{<:RiemannianFunction})

The JuMP function of a RiemannianFunction for any AbstractModel is that function itself.

JuMP.jump_function_type — Method

JuMP.jump_function_type(::JuMP.AbstractModel, F::Type{<:RiemannianFunction})

The JuMP function type of a function of type RiemannianFunction for any AbstractModel is that function type itself

JuMP.set_objective_function — Method

JuMP.set_objective_function(model::JuMP.Model, obj::Manopt.AbstractManifoldObjective)

Set the objective function of a JuMP.Model model to an AbstractManifoldObjective obj. This allows to use @objective with an objective from Manopt.jl.

Manopt.JuMP_ManifoldSet — Method

JuMP_ManifoldSet(M::ManifoldsBase.AbstractManifold) = ManifoldSet(M)

Create a ManifoldSet.

Manopt.JuMP_Optimizer — Method

Manopt.JuMP_Optimizer()

Represent a solver from Manopt.jl within the MathOptInterface (MOI) framework.

See ManoptOptimizer for the fields and their meaning.

ManoptJuMPExt._reshape_vector! — Method

_reshape_vector!(res::Array{T,N}, vec::Vector{T}, ::ManifoldArrayShape{N}) where {T,N}

Inplace version of res = JuMP.reshape_vector(vec, shape).

ManoptJuMPExt._vectorize! — Method

_vectorize!(res::Vector{T}, array::Array{T,N}, shape::ManifoldArrayShape{N}) where {T,N}

Inplace version of res = JuMP.vectorize(array, shape).

ManoptJuMPExt._zero — Method

_zero(shape::ManifoldArrayShape)

Return a zero element of the shape shape.

MathOptInterface.Utilities.map_indices — Method

MOI.Utilities.map_indices(index_map::Function, func::RiemannianFunction)

The original docstring states something about substituting some variable indices by their index map variants. On a RiemannianFunction there is nothing to substitute,

MathOptInterface.add_constrained_variables — Method

MOI.add_constrained_variables(model::ManoptOptimizer, set::ManifoldSet)

Add MOI.dimension(set) variables constrained in set and return the list of variable indices that can be used to reference them as well a constraint index for the constraint enforcing the membership of the variables in the ManifoldSet set.

MathOptInterface.copy_to — Method

MOI.copy_to(dest::ManoptOptimizer, src::MOI.ModelLike)

Because supports_incremental_interface(dest) is true, this simply uses MOI.Utilities.default_copy_to and copies the variables with MOI.add_constrained_variables and the objective sense with MOI.set.

MathOptInterface.dimension — Method

MOI.dimension(set::ManifoldSet)

Return the representation side of points on the (vectorized in representation) manifold. As the MOI variables are real, this means if the representation_size yields (in product) n, this refers to the vectorized point / tangent vector from (a subset of $ℝ^n$).

MathOptInterface.empty! — Method

MOI.empty!(model::ManoptJuMPExt.Optimizer)

Clear all model data from model but keep the options set.

MathOptInterface.get — Method

MOI.get(::ManoptOptimizer, ::MOI.DualStatus)

Returns MOI.NO_SOLUTION indicating that there is no dual solution available.

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, ::MOI.NumberOfVariables)

Return the number of variables added in the model, this corresponds to the MOI.dimension of the ManifoldSet.

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, ::MOI.ObjectiveSense)

Return the objective sense, defaults to MOI.FEASIBILITY_SENSE if no sense has already been set.

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, attr::MOI.ObjectiveValue)

Return the value of the objective function evaluated at the solution.

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, ::MOI.PrimalStatus)

Return MOI.NO_SOLUTION if optimize! hasn't been called yet and MOI.FEASIBLE_POINT otherwise indicating that a solution is available to query with MOI.VariablePrimalStart.

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)

Return last value set by MOI.set(model, attr, value).

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, ::MOI.RawStatusString)

Return a String containing Manopt.get_reason without the ending newline character.

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, ::MOI.ResultCount)

Return 0 if optimize! hasn't been called yet and 1 otherwise indicating that one solution is available.

MathOptInterface.get — Method

MOI.get(::ManoptOptimizer, ::MOI.SolverName)

Return the name of the Optimizer with the value of the descent_state_type option.

MathOptInterface.get — Method

MOI.get(::ManoptOptimizer, ::MOI.SolverVersion)

Return the version of the Manopt solver, it corresponds to the version of Manopt.jl.

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, ::MOI.ResultCount)

Return MOI.OPTIMIZE_NOT_CALLED if optimize! hasn't been called yet and MOI.LOCALLY_SOLVED otherwise indicating that the solver has solved the problem to local optimality the value of MOI.RawStatusString for more details on why the solver stopped.

MathOptInterface.get — Method

MOI.get(model::ManoptOptimizer, attr::MOI.VariablePrimal, vi::MOI.VariableIndex)

Return the value of the solution for the variable of index vi.

MathOptInterface.is_valid — Method

MOI.is_valid(model::ManoptOptimizer, vi::MOI.VariableIndex)

Return whether vi is a valid variable index.

MathOptInterface.set — Method

MOI.set(model::ManoptOptimizer, ::MOI.ObjectiveFunction{F}, func::F) where {F}

Set the objective function as func for model.

MathOptInterface.set — Method

MOI.set(model::ManoptOptimizer, ::MOI.ObjectiveSense, sense::MOI.OptimizationSense)

Modify the objective sense to either MOI.MAX_SENSE, MOI.MIN_SENSE or MOI.FEASIBILITY_SENSE.

MathOptInterface.set — Method

MOI.get(model::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)

Set the value for the keyword argument attr.name to give for the constructor model.options[DESCENT_STATE_TYPE].

MathOptInterface.set — Method

function MOI.set(
    model::ManoptOptimizer,
    ::MOI.VariablePrimalStart,
    vi::MOI.VariableIndex,
    value::Union{Real,Nothing},
)

Set the starting value of the variable of index vi to value. Note that if value is nothing then it essentially unset any previous starting values set and hence MOI.optimize! unless another starting value is set.

MathOptInterface.supports — Method

MOI.supports(::ManoptOptimizer, attr::MOI.RawOptimizerAttribute)

Return a Bool indicating whether attr.name is a valid option name for Manopt.

MathOptInterface.supports — Method

MOI.supports(::Manopt.JuMP_Optimizer, attr::MOI.RawOptimizerAttribute)

Return true indicating that Manopt.JuMP_Optimizer supports starting values for the variables.

MathOptInterface.supports — Method

MOI.supports(::ManoptOptimizer, ::Union{MOI.ObjectiveSense,MOI.ObjectiveFunction})

Return true indicating that Optimizer supports being set the objective sense (that is, min, max or feasibility) and the objective function.

MathOptInterface.supports_add_constrained_variables — Method

MOI.supports_add_constrained_variables(::JuMP_Optimizer, ::Type{<:ManifoldSet})

Return true indicating that Manopt.JuMP_Optimizer support optimization on variables constrained to belong in a vectorized manifold ManifoldSet.

MathOptInterface.supports_incremental_interface — Method

MOI.supports_incremental_interface(::JuMP_Optimizer)

Return true indicating that Manopt.JuMP_Optimizer implements MOI.add_constrained_variables and MOI.set for MOI.ObjectiveFunction so it can be used with JuMP.direct_model and does not require a MOI.Utilities.CachingOptimizer. See MOI.supports_incremental_interface.