cost functions

Cost Functions

The following cost functions are available

Manopt.costIntrICTV12 — Method.

costIntrICTV12(M,f,u,v,α,β)

computes the intrinsic infimal convolution model, where the addition is replaced by a mid point approach and the two functions involved are costTV2 and costTV. The model reads

\[E(u,v) = \frac{1}{2}\sum_{i\in\mathcal G} d_{\mathcal M}\bigl(g(\frac{1}{2},v_i,w_i),f_i\bigr) +\alpha\bigl( \beta\mathrm{TV}(v) + (1-\beta)\mathrm{TV}_2(w) \bigr).\]

source

Manopt.costL2TV — Method.

costL2TV(M,f,α,x)

compute the $\ell^2$-TV functional on the Power manifold M for given (fixed) data f (on M), a nonnegative weight α, and evaluated at x (on M), i.e.

\[E(x) = d_{\mathcal M}^2(f,x) + \alpha \operatorname{TV}(x)\]

See also

costTV

source

Manopt.costL2TV2 — Method.

costL2TV2(M,f,β,x)

compute the $\ell^2$-TV2 functional on the Power manifold M for given data f, nonnegative parameter β, and evaluated at x, i.e.

\[E(x) = d_{\mathcal M}^2(f,x) + \beta\operatorname{TV}_2(x)\]

See also

costTV2

source

Manopt.costL2TVTV2 — Method.

costL2TVTV2(M,f,α,β,x)

compute the $\ell^2$-TV-TV2 functional on the Power manifold M for given (fixed) data f (on M), nonnegative weight α, β, and evaluated at x (on M), i.e.

\[E(x) = d_{\mathcal M}^2(f,x) + \alpha\operatorname{TV}(x) + \beta\operatorname{TV}_2(x)\]

See also

costTV, costTV2

source

Manopt.costTV — Function.

costTV(M,x [,p=2,q=1])

compute the $\operatorname{TV}^p$ functional for data xon the Power manifold M, i.e. $\mathcal M = \mathcal N^n$, where $n\in\mathbb N^k$ denotes the dimensions of the data x. Let $\mathcal I_i$ denote the forward neighbors, i.e. with $\mathcal G$ as all indices from $\mathbf{1}\in\mathbb N^k$ to $n$ we have $\mathcal I_i = \{i+e_j, j=1,\ldots,k\}\cap \mathcal G$. The formula reads

\[E^q(x) = \sum_{i\in\mathcal G} \bigl( \sum_{j\in \mathcal I_i} d^p_{\mathcal M}(x_i,x_j) \bigr)^{q/p}.\]

See also

gradTV, proxTV

source

Manopt.costTV — Method.

costTV(M,x,p)

compute the $\operatorname{TV}^p$ functional for a tuple pT of MPoints on a Manifold M, i.e.

\[E(x_1,x_2) = d_{\mathcal M}^p(x_1,x_2), \quad x_1,x_2\in\mathcal M\]

See also

gradTV, proxTV

source

Manopt.costTV2 — Function.

costTV2(M,x [,p=1])

compute the $\operatorname{TV}_2^p$ functional for data x on the Power manifoldmanifold M, i.e. $\mathcal M = \mathcal N^n$, where $n\in\mathbb N^k$ denotes the dimensions of the data x. Let $\mathcal I_i^{\pm}$ denote the forward and backward neighbors, respectively, i.e. with $\mathcal G$ as all indices from $\mathbf{1}\in\mathbb N^k$ to $n$ we have $\mathcal I^\pm_i = \{i\pm e_j, j=1,\ldots,k\}\cap \mathcal I$. The formula then reads

\[E(x) = \sum_{i\in\mathcal I,\ j_1\in \mathcal I^+_i,\ j_2\in \mathcal I^-_i} d^p_{\mathcal M}(c_i(x_{j_1},x_{j_2}), x_i),\]

where $c_i(\cdot,\cdot)$ denotes the mid point between its two arguments that is nearest to $x_i$.

See also

gradTV2, proxTV2

source

Manopt.costTV2 — Method.

costTV2(M,(x1,x2,x3) [,p=1])

compute the $\operatorname{TV}_2^p$ functional for the 3-tuple of points (x1,x2,x3)on the Manifold M. Denote by

\[ \mathcal C = \bigl\{ c\in \mathcal M \ |\ g(\tfrac{1}{2};x_1,x_3) \text{ for some geodesic }g\bigr\}\]

the set of mid points between $x_1$ and $x_3$. Then the functionr reads

\[d_2^p(x_1,x_2,x_3) = \min_{c\in\mathcal C} d_{\mathcal M}(c,x_2).\]

See also

gradTV2, proxTV2

source