Jacobi Fields

A smooth tangent vector field $J\colon [0,1] \to T\mathcal M$ along a geodesic $g(\cdot;x,y)$ is called Jacobi field if it fulfills the ODE

\[\displaystyle 0 = \frac{D}{dt}J + R(J,\dot g)\dot g,\]

where $R$ is the Riemannian curvature tensor. Such Jacobi fields can be used to derive closed forms for the exponential map, the logarithmic map and the geodesic, all of them with respect to both arguments: Let $F\colon\mathcal N \to \mathcal M$ be given (for the $\exp_x\cdot$ we have $\mathcal N = T_x\mathcal M$, otherwise $\mathcal N=\mathcal M$) and denote by $\Xi_1,\ldots,\Xi_d$ an orthonormal frame along $g(\cdot;x,y)$ that diagonalizes the curvature tensor with corresponding eigenvalues $\kappa_1,\ldots,\kappa_d$. Note that on symmetric manifolds such a frame always exists.

Then $DF(x)[\eta] = \sum_{k=1}^d \langle \eta,\Xi_k(0)\rangle_x\beta(\kappa_k)\Xi_k(T)$ holds, where $T$ also depends on the function $F$ as the weights $\beta$. The values stem from solving the corresponding system of (decoupled) ODEs.

Note that in different references some factors might be a little different, for example when using unit speed geodesics.

The following weights functions are available

Manopt.adjoint_Jacobi_field — Function

Y = adjoint_Jacobi_field(M, p, q, t, X, β)

Compute the AdjointJacobiField $J$ along the geodesic $γ_{p,q}$ on the manifold $\mathcal M$ with initial conditions (depending on the application) $X ∈ T_{γ_{p,q}(t)}\mathcal M$ and weights $β$. The result is a vector $Y ∈ T_p\mathcal M$. The main difference to jacobi_field is the, that the input X and the output Y switched tangent spaces. For detais see jacobi_field

Manopt.jacobi_field — Function

Y = jacobi_field(M, p, q, t, X, β)

compute the Jacobi jield $J$ along the geodesic $γ_{p,q}$ on the manifold $\mathcal M$ with initial conditions (depending on the application) $X ∈ T_p\mathcal M$ and weights $β$. The result is a tangent vector Y from $T_{γ_{p,q}(t)}\mathcal M$.

See also

adjoint_Jacobi_field

Manopt.βdifferential_exp_argument — Method

βdifferential_exp_argument(κ,t,d)

weights for the jacobi_field corresponding to the differential of the geodesic with respect to its start point $D_X \exp_p X[Y]$. They are

\[$\beta(\kappa) = \begin{cases} \frac{\sinh(d\sqrt{-\kappa})}{d\sqrt{-\kappa}}&\text{ if }\kappa < 0,\\ 1 & \text{ if } \kappa = 0,\\ \frac{\sin(d\sqrt{\kappa})}{\sqrt{d\kappa}}&\text{ if }\kappa > 0. \end{cases}\]

See also

differential_exp_argument, jacobi_field

Manopt.βdifferential_exp_basepoint — Method

βdifferential_exp_basepoint(κ,t,d)

weights for the jacobi_field corresponding to the differential of the geodesic with respect to its start point $D_p \exp_p X [Y]$. They are

\[\beta(\kappa) = \begin{cases} \cosh(\sqrt{-\kappa})&\text{ if }\kappa < 0,\\ 1 & \text{ if } \kappa = 0,\\ \cos(\sqrt{\kappa}) &\text{ if }\kappa > 0. \end{cases}\]

See also

differential_exp_basepoint, jacobi_field

Manopt.βdifferential_geodesic_startpoint — Method

βdifferential_geodesic_startpoint(κ,t,d)

weights for the jacobi_field corresponding to the differential of the geodesic with respect to its start point $D_x g(t;p,q)[X]$. They are

\[\beta(\kappa) = \begin{cases} \frac{\sinh(d(1-t)\sqrt{-\kappa})}{\sinh(d\sqrt{-\kappa})} &\text{ if }\kappa < 0,\\ 1-t & \text{ if } \kappa = 0,\\ \frac{\sin((1-t)d\sqrt{\kappa})}{\sinh(d\sqrt{\kappa})} &\text{ if }\kappa > 0. \end{cases}\]

Due to a symmetry agrument, these are also used to compute $D_q g(t; p,q)[\eta]$

See also

differential_geodesic_endpoint, differential_geodesic_startpoint, jacobi_field

Manopt.βdifferential_log_argument — Method

βdifferential_log_argument(κ,t,d)

weights for the JacobiField corresponding to the differential of the logarithmic map with respect to its argument $D_q \log_p q[X]$. They are

\[\beta(\kappa) = \begin{cases} \frac{ d\sqrt{-\kappa} }{\sinh(d\sqrt{-\kappa})}&\text{ if }\kappa < 0,\\ 1 & \text{ if } \kappa = 0,\\ \frac{ d\sqrt{\kappa} }{\sin(d\sqrt{\kappa})}&\text{ if }\kappa > 0. \end{cases}\]

See also

differential_log_basepoint, jacobi_field

Manopt.βdifferential_log_basepoint — Method

βdifferential_log_basepoint(κ,t,d)

weights for the jacobi_field corresponding to the differential of the geodesic with respect to its start point $D_p \log_p q[X]$. They are

\[\beta(\kappa) = \begin{cases} -\sqrt{-\kappa}d\frac{\cosh(d\sqrt{-\kappa})}{\sinh(d\sqrt{-\kappa})}&\text{ if }\kappa < 0,\\ -1 & \text{ if } \kappa = 0,\\ -\sqrt{\kappa}d\frac{\cos(d\sqrt{\kappa})}{\sin(d\sqrt{\kappa})}&\text{ if }\kappa > 0. \end{cases}\]

See also

differential_log_argument, differential_log_argument, jacobi_field