The Euclidean Space mathbb R^n

The $n$-dimensional Euclidean Space $\mathbb R^n$

The Euclidean space serves as a fallback to standard methods or as a part of a Product. Let's start with the following instances of the abstract types Manifold, MPoint, and TVector.

Manopt.Euclidean — Type.

Euclidean <: Manifold

The manifold $\mathcal M = \mathbb R^n$ of the $n$-dimensional Euclidean vector space. We employ the notation $\langle\cdot,\cdot\rangle$ for the inner product and $\lVert\cdot\rVert_2$ for its induced norm.

Abbreviation

Rn

Constructor

Euclidean(n)

construct the n-dimensional Euclidean space $\mathbb R^n$.

source

Manopt.RnPoint — Type.

RnPoint <: MPoint

the point $x\in\mathcal M$ for $\mathcal M=\mathbb R^n$ represented by an $n$-dimensional Vector{T}, where T <: AbstractFloat.

source

Manopt.RnTVector — Type.

RnTVector <: TVector

the tangent vector $\xi \in T_x\mathcal M$ for $\mathcal M=\mathbb R^n$ represented by an $n$-dimensional Vector{T}, where T <: AbstractFloat.

source

Functions

Base.exp — Method.

exp(M,x,ξ[, t=1.0])

compute the exponential map on the Euclidean manifold M, i.e. $x+t*\xi$, where the scaling parameter t is optional.

source

Base.log — Method.

log(M,x,y)

computes the logarithmic map on the Euclidean manifold M, i.e. $y-x$.

source

LinearAlgebra.dot — Method.

dot(M,x,ξ,ν)

Computes the Euclidean inner product of ξ and ν, i.e. $\langle\xi,\nu\rangle = \displaystyle\sum_{k=1}^n \xi_k\nu_k$.

source

Manopt.distance — Method.

distance(M,x,y)

compute the Euclidean distance $\lVert x - y\rVert$

source

Manopt.injectivityRadius — Method.

injectivityRadius(M)

return the injectivity radius of the Euclidean manifold M$=\mathbb R^n$.

source

Manopt.manifoldDimension — Method.

manifoldDimension(M)

return the manifold dimension of the Euclidean manifold M, i.e. the length of the vectors stored in M.dimension, i.e. $n$.

source

Manopt.manifoldDimension — Method.

manifoldDimension(x)

return the manifold dimension of the RnPoint x, i.e. $n$.

source

Manopt.parallelTransport — Method.

parallelTransport(M,x,y,ξ)

compute the parallel transport the Euclidean manifold M, which is the identity.

source

Manopt.project — Method.

project(M,x,v)

project a $n$-dimensional Vector{T} v on the tangent space of the RnPoint{T} x. Since the tangent space is identical to the $\mathbb R^n$, the mapping can be realized with the identity, i.e. $\operatorname{project}(M,x,v) = v$.

source

Manopt.randomMPoint — Function.

randomMPoint(M[,T=Float64])

generate a random point on the Euclidean manifold M, where the optional parameter determines the type of the entries of the resulting RnPoint.

source

Manopt.tangentONB — Method.

(Ξ,κ) = tangentONB(M,x,y)

compute an ONB within the tangent space $T_x\mathcal M$ at the MPoint on the Euclidean manifold M, such that $\xi=\log_xy$ is the first vector and compute the eigenvalues of the curvature tensor $R(\Xi,\dot g)\dot g$, where $g=g_{x,\xi}$ is the geodesic with $g(0)=x$, $\dot g(0) = \xi$, i.e. $\kappa_1$ corresponding to $\Xi_1=\xi$ is zero.

See also

jacobiField, adjointJacobiField.

source

Manopt.typicalDistance — Method.

typicalDistance(M)

returns the typical distance on the Euclidean manifold M: $\sqrt{n}$.

source

Manopt.validateMPoint — Method.

validateMPoint(M,x)

Checks that a RnPoint x has a valid value for a point on the Euclidean manifold M$=\mathbb R^n$, which is the case if the dimensions fit.

source

Manopt.validateTVector — Method.

validateTVector(M,x,ξ)

Checks, that the RnTVector ξ is a valid tangent vector in the tangent space of the RnPoint x ont the Euclidean manifold M, which is always the case as long as their vector dimensions agree.

source

Manopt.zeroTVector — Method.

ξ = zeroTVector(M,x)

returns a zero vector in the tangent space $T_x\mathcal M$ of the RnPoint $x\in\mathbb R^n$ on the Euclidean manifold M.

source