Optimization with box domains and products of manifolds and boxes

A Hyperrectangle is, in general, not a manifold but a manifold with corners, thus handling it as a domain in optimization requires special attention. For simple methods like gradient descent using projected gradient and a stopping criterion involving StopWhenProjectedNegativeGradientNormLess may be sufficient, however methods that approximate the Hessian can benefit from a more advanced approach. The core idea is considering a piecewise quadratic approximation of the objective along the descent direction, and selecting the generalized Cauchy point – its minimizer. The points at which the approximation might not be differentiable correspond to hitting new boundaries along the initially selected descent direction. Then, we can perform standard line search between the initial iterate and the generalized Cauchy point.

Public types and method

Manopt.QuasiNewtonLimitedMemoryBoxDirectionUpdate — Type

mutable struct LimitedMemoryHessianApproximation end

An approximation of Hessian of a scalar function of the form $B_0 = θ I$, $B_{k+1} = B_k - W_k M_k W_k^{\mathrm{T}}$, where $\theta > 0$ is an initial scaling guess. Matrix $M_k = \left(\begin{smallmatrix}M_{11} & M_{21}^{\mathrm{T}}\\ M_{21} & M_{22}\end{smallmatrix}\right)$ is stored using its blocks. Blocks $W_k$ are (implicitly) composed from memory_y and memory_s.

Initial scale $\theta$ is stored in the field initial_scale but if the memory isn't empty, the current scale is set to squared norm of $s_k$ divided by inner product of $s_k$ and $y_k$ where $k$ is the oldest index for which the denominator is not equal to 0.

See [BNS94] for details.

Internal types and method

Manopt.init_updater! — Function

init_updater!(::AbstractManifold, fpfpp_upd::AbstractFPFPPUpdater, p, d, ha::AbstractQuasiNewtonDirectionUpdate)

Method for initialization of AbstractFPFPPUpdater fpfpp_upd just before the loop that examines subsequent intervals for GCP. By default it does nothing.

Manopt.hess_val — Function

hess_val(gh::QuasiNewtonLimitedMemoryBoxDirectionUpdate, M::AbstractManifold, p, X)

Compute $⟨X, B X⟩$, where $B$ is the (1, 1)-Hessian represented by gh.

Manopt.AbstractFPFPPUpdater — Type

abstract type AbstractFPFPPUpdater end

Abstract type for methods that calculate f' and f'' in the GCP calculation in subsequent line segments in GCPFinder.

Manopt.GenericFPFPPUpdater — Type

struct GenericFPFPPUpdater <: AbstractFPFPPUpdater end

Generic f' and f'' calculation that only relies on hess_val_eb but is relatively slow for high-dimensional domains.

Manopt.get_bounds_index — Function

get_bounds_index(::AbstractManifold)

Get the bound indices of manifold M. Standard manifolds don't have bounds, so Base.OneTo(1) is returned.

Manopt.requires_gcp — Function

requires_gcp(M::AbstractManifold)

Return true if M is a Hyperrectangle-like manifold with corners, or a product of it with a standard manifold. Otherwise return false.

Manopt.find_gcp_direction! — Function

find_gcp_direction!(gcp::GCPFinder, d_out, p, d, X)

Find generalized Cauchy point looking from point p in direction d and save the tangent vector pointing at it to d_out. Gradient of the objective at p is X.

The function returns

:found_limited if the point was found and we can perform a step of length at most 1 in direction d_out afterwards,

:found_unlimited if the point was found and we can perform a step of length at most max_stepsize(M, p) in direction d_out afterwards,

:not_found if the search cannot be performed in direction d.

Manopt.hess_val_eb — Function

hess_val_eb(gh::QuasiNewtonLimitedMemoryBoxDirectionUpdate, M::AbstractManifold, p, b)

Compute $⟨X, B X⟩$, where $B$ is the (1, 1)-Hessian represented by gh, and X is the unit vector along index b.

hess_val_eb(gh::QuasiNewtonLimitedMemoryBoxDirectionUpdate, M::AbstractManifold, p, b, Y)

Compute $⟨X, B Y⟩$, where $B$ is the (1, 1)-Hessian represented by gh, where X is the unit vector pointing at index b.

Manopt.LimitedMemoryFPFPPUpdater — Type

struct LimitedMemoryFPFPPUpdater{TV <: AbstractVector} <: AbstractFPFPPUpdater

f' and f'' calculation that is optimized for QuasiNewtonLimitedMemoryBoxDirectionUpdate. It relies on a specific Hessian structure.

Manopt.get_bound_t — Function

get_bound_t(M::AbstractManifold, x, d, i)

Get the upper bound on moving in direction d from point p on manifold M, for the bound index i.