Gradient Descent

Manopt.gradient_descent — Function

gradient_descent(M, F, ∇F, x)

perform a gradientdescent x{k+1} = \mathrm{retr}{xk} sk\nabla f(xk)$ with different choices of $s_k$ available (see stepsize option below).

Input

M – a manifold $\mathcal M$

F – a cost function $F\colon\mathcal M\to\mathbb R$ to minimize

∇F – the gradient $\nabla F\colon\mathcal M\to T\mathcal M$ of F

x – an initial value $x ∈ \mathcal M$

Optional

stepsize – (ConstantStepsize(1.)) specify a Stepsize functor.

retraction – (exp) a retraction(M,x,ξ) to use.

stopping_criterion – ([StopWhenAny](@ref)([StopAfterIteration](@ref)(200), [StopWhenGradientNormLess](@ref)(10.0^-8))) a functor inheriting from [StoppingCriterion`](@ref) indicating when to stop.

return_options – (false) – if activated, the extended result, i.e. the complete Options are returned. This can be used to access recorded values. If set to false (default) just the optimal value xOpt if returned

... and the ones that are passed to decorate_options for decorators.

Output

xOpt – the resulting (approximately critical) point of gradientDescent

OR

options - the options returned by the solver (see return_options)

Options

Manopt.GradientDescentOptions — Type

GradientDescentOptions{P,T} <: Options

Describes a Gradient based descent algorithm, with

Fields

a default value is given in brackets if a parameter can be left out in initialization.

x0 – an a point (of type P) on a manifold as starting point

stopping_criterion – (StopAfterIteration(100)) a StoppingCriterion

stepsize – (ConstantStepsize(1.))a Stepsize

retraction_method – (ExponentialRetraction()) the rectraction to use, defaults to the exponential map

Constructor

GradientDescentOptions(x, stop, s [, retr=ExponentialRetraction()])

construct a Gradient Descent Option with the fields and defaults as above

See also

gradient_descent, GradientProblem

Debug Actions

Manopt.DebugGradient — Type

DebugGradient <: DebugAction

debug for the gradient evaluated at the current iterate

Constructors

DebugGradient([long=false,p=print])

display the short (false) or long (true) default text for the gradient.

DebugGradient(prefix[, p=print])

display the a prefix in front of the gradient.

Manopt.DebugGradientNorm — Type

DebugGradientNorm <: DebugAction

debug for gradient evaluated at the current iterate.

Constructors

DebugGradientNorm([long=false,p=print])

display the short (false) or long (true) default text for the gradient norm.

DebugGradientNorm(prefix[, p=print])

display the a prefix in front of the gradientnorm.

Manopt.DebugStepsize — Type

DebugStepsize <: DebugAction

debug for the current step size.

Constructors

DebugStepsize([long=false,p=print])

display the short (false) or long (true) default text for the step size.

DebugStepsize(prefix[, p=print])

display the a prefix in front of the step size.

Record Actions

Manopt.RecordGradient — Type

RecordGradient <: RecordAction

record the gradient evaluated at the current iterate

Constructors

RecordGradient(ξ)

initialize the RecordAction to the corresponding type of the tangent vector.

Manopt.RecordGradientNorm — Type

RecordGradientNorm <: RecordAction

record the norm of the current gradient

Manopt.RecordStepsize — Type

RecordStepsize <: RecordAction

record the step size