cost functions

Cost Functions

The following cost functions are available

Manopt.costIntrICTV12 — Method.

costIntrICTV12(M,f,u,v,α,β) Computes the intrinsic infimal convolution model using the mid point model of second order differences, see costTV2, i.e. for some $f\in\mathcal M$ on a [Power] manifold $\mathcal M$ this function computes

\[E_{\mathrm{IC}}^{\mathrm{int}}(u,v) = \frac{1}{2}\sum_{i\in\mathcal G} d_{\mathcal M}(g(\frac{1}{2},v_i,w_i),f_i) + \alpha \bigl( \beta\mathrm{TV}(v) + (1-\beta)\mathrm{TV}_2(w) \bigr),\]

where both total variations refer to the intrinsic ones, costTV and costTV2, respectively

source

Manopt.costL2TV — Method.

costL2TV(M,f,α,x)

compute the L2-TV functional for given data f and nonnegative parameter α, i.e.

$ E(x) = d_{\mathcal M}^2(f,x) + \alpha \operatorname{TV}(x) $

where $\mathcal M = \mathcal N^n$ is a PowerManifold, i.e. $\mathcal N$ is a Manifold for all entries of an array of size $n\in\mathbb N^k, k>0$.

See also

costTV

source

Manopt.costL2TV2 — Method.

costL2TV2(M,f,β,x)

compute the L2-TV+TV2 functional for given data f and nonnegative parameter β, i.e.

$ E(x) = d{\mathcal M}^2(f,x) + \beta\operatorname{TV}2(x) $

where $\mathcal M = \mathcal N^n$ is a PowerManifold, i.e. $\mathcal N$ is a Manifold for all entries of an array of size $n\in\mathbb N^k, k>0$.

See also

costTV, costTV2

source

Manopt.costL2TVplusTV2 — Method.

costL2TVplusTV2(M,f,α,β,x)

compute the L2-TV+TV2 functional for given data f and nonnegative parameters α and β, i.e.

$ E(x) = d{\mathcal M}^2(f,x) + \alpha\operatorname{TV}(x) + \beta\operatorname{TV}2(x) $

where $\mathcal M = \mathcal N^n$ is a PowerManifold, i.e. $\mathcal N$ is a Manifold for all entries of an array of size $n\in\mathbb N^k, k>0$.

See also

costTV, costTV2

source

Manopt.costTV — Function.

costTV(M,x[,p=2,q=1])

compute the $\operatorname{TV}^p$ functional for signal, image or dataset x on the PowerManifold M, i.e. $\mathcal M = \mathcal N^n$, where $n\in\mathbb N^k$ denotes the dimensions of the data. Denoting by $\mathcal I$ all indices from $\mathbf{1}\in\mathbb N^k$ to $n$ and $\mathcal I^+_i = \{i+e_j, j=1,\ldots,k\}\cap \mathcal I$ its forward neighbors, this function computes

$ E^q(x) = \sum{i\in\mathcal I} \Bigl( \sum{j\in \mathcal I^+i} d^p{\mathcal M}^p(xi,x_j) \bigr)^{q/p}, \quad x\in \mathcal M$

See also

gradTV, proxTV

source

Manopt.costTV — Method.

costTV(M,x,p)

compute the $\operatorname{TV}^p$ functional for a tuple pT of MPoint data points on a Manifold M, i.e.

$ E(x1,x2) = d{\mathcal M}^p(x1,x2), \quad x1,x_2\in\mathcal M $

See also

gradTV, proxTV

source

Manopt.costTV2 — Function.

costTV2(M,x[p=1])

compute the $\operatorname{TV}^p$ functional for signal, image or dataset x on the PowerManifold M, i.e. $\mathcal M = \mathcal N^n$, where $n\in\mathbb N^k$ denotes the dimensions of the data. Denoting by $\mathcal I$ all indices from $\mathbf{1}\in\mathbb N^k$ to $n$ and $\mathcal I^\pm_i = \{\pm e_j, j=1,\ldots,k\}\cap \mathcal I$ its forward and backward neighbors, respectively, this function computes

$ E(x) = \sum{i\in\mathcal I, j1\in \mathcal I^+i,j2\in \mathcal I^-i} d^p{\mathcal M}^p(c(x{j1},x{j2}), x_i), \quad x\in \mathcal M,$

where $c(\cdot,\cdot)$ denotes the mid point between its two arguments nearest to $x_i$.

See also

gradTV2, proxTV2

source

Manopt.costTV2 — Method.

costTV2(M,(x1,x2,x3),[p=1])

compute the $\operatorname{TV}_2^p$ functional for the 3-tuple of points (x1,x2,x3)on the [Manifold](@ref)M`, denote by $\mathcal C = \{ c\in \mathcal M | g(\frac{1}{2};x_1,x_3) \text{ for some geodesic }g\}$ the set of mid points between the first and third point. Then this function computes

\[d_2^p(x_1,x_2,x_3) = \min_{c\in\mathcal C} d_{\mathcal M}(c,x_2).\]

See also

gradTV2, proxTV2

source