Differentials

Differentials

Manopt.DforwardLogs — Method.

ν = DforwardLogs(M,x,ξ)

compute the differenital of forwardLogs $F$ on the Power manifold M at x and direction ξ , in the power manifold array, the differential of the function

\[F_i(x) = \sum_{j\in\mathcal I_i} \log_{x_i} x_j$, \quad i \in \mathcal G,\]

where $\mathcal G$ is the set of indices of the Power manifold M and $\mathcal I_i$ denotes the forward neighbors of $i$.

Input

M – a Power manifold

x – a PowPoint.

ξ – a PowTVector.

Ouput

ν – resulting tangent vector in $T_x\mathcal N$ representing the differentials of the logs, where $\mathcal N$ is thw power manifold with the number of dimensions added to size(x).

source

Manopt.DxExp — Method.

DxExp(M,x,ξ,η)

computes $D_x\exp_x\xi[\eta]$.

See also

DξExp, jacobiField

source

Manopt.DxGeo — Method.

DxGeo(M,x,y,t,η)

computes $D_xg(t;x,y)[\eta]$.

See also

DyGeo, jacobiField

source

Manopt.DxLog — Method.

DxLog(M,x,y,η)

computes $D_xlog_xy[\eta]$.

See also

DyLog, jacobiField

source

Manopt.DyGeo — Method.

DyGeo(M,x,y,t,η)

computes $D_yg(t;x,y)[\eta]$.

See also

DxGeo, jacobiField

source

Manopt.DyLog — Method.

DyLog(M,x,y,η)

computes $D_ylog_xy[\eta]$.

See also

DxLog, jacobiField

source

Manopt.DξExp — Method.

DξExp(M,x,ξ,η)

computes $D_\xi\exp_x\xi[\eta]$. Note that $\xi\in T_\xi(T_x\mathcal M) = T_x\mathcal M$ is still a tangent vector.

See also

DxExp, jacobiField

source