Gradients

Gradients

For a function $f\colon\mathcal M\to\mathbb R$ the Riemannian gradient $\nabla f(x)$ at $x\in\mathcal M$ is given by the unique tangent vector fulfilling

\[\langle \nabla f(x), \xi\rangle_x = D_xf[\xi],\quad \forall \xi \in T_x\mathcal M,\]

where $D_xf[\xi]$ denotes the differential of $f$ at $x$ with respect to the tangent direction (vector) $\xi$ or in other words the directional derivative.

This page collects the available gradients.

Manopt.forwardLogs — Method.

ξ = forwardLogs(M,x)

Compute the forward logs (generalizing forward differences) orrucirng, in the power manifold array, the function

\[F(x) = \sum_{i}\sum_{j\in\mathcal N_i} \log_{x_i} x_j\]

where $i$ runs over all indices of the Power manifold M and $\mathcal N_i$ denotes the forward neighbors of $i$.

Input

M : a Power manifold

x : a PowPoint.

Ouput

ξ : resulting tangent vector in $T_x\mathcal M$ representing the logs, where $\mathcal N$ is thw power manifold with the number of dimensions added to size(x).

source

Manopt.gradDistance — Function.

gradDistance(M,y,x[, p=2])

compute the (sub)gradient of the distance (squared)

\[f(x) = \frac{1}{2} d_{\mathcal M}(x,y)\]

to a fixed MPointy on the Manifold M.

Optional

p – (2) the exponent of the distance, i.e. the default is the squared distance

source

Manopt.gradIntrICTV12 — Method.

∇u,⁠∇v = gradIntrICTV12(M,f,u,v,α,β) Computes (sub)gradient of the intrinsic infimal convolution model using the mid point model of second order differences, see costTV2, i.e. for some $f\in\mathcal M$ on a [Power] manifold $\mathcal M$ this function computes the (sub) gradient of

\[E_{\mathrm{IC}}^{\mathrm{int}}(u,v) = \frac{1}{2}\sum_{i\in\mathcal G} d_{\mathcal M}(g(\frac{1}{2},v_i,w_i),f_i) + \alpha \bigl( \beta\mathrm{TV}(v) + (1-\beta)\mathrm{TV}_2(w) \bigr),\]

where both total variations refer to the intrinsic ones, gradTV and gradTV2, respectively.

source

Manopt.gradTV — Function.

gradTV(M,(x,y),[p=1])

computes the (sub) gradient of $\frac{1}{p}d^p_{\mathcal M}(x,y)$ with respect to both $x$ and $y$.

source

Manopt.gradTV — Function.

ξ = gradTV(M,λ,x,[p])

Compute the (sub)gradient $\partial F$ of all forward differences orrucirng, in the power manifold array, i.e. of the function

\[F(x) = \sum_{i}\sum_{j\in\mathcal N_i} d^p(x_i,x_j)\]

where $i$ runs over all indices of the Power manifold M and $\mathcal N_i$ denotes the forward neighbors of $i$.

Input

M : a Power manifold

x : a PowPoint.

Ouput

ξ : resulting tangent vector in $T_x\mathcal M$.

source

Manopt.gradTV2 — Function.

gradTV2(M,x,p)

computes the (sub) gradient of $\frac{1}{p}d_2^p_{\mathcal M}(x_1,x_2,x_3)$ with respect to all $x_1,x_2,x_3$ occuring along any array dimension in the PowPoint x, where M is the corresponding Power manifold.

source

Manopt.gradTV2 — Function.

gradTV2(M,(x,y,z),p)

computes the (sub) gradient of $\frac{1}{p}d_2^p_{\mathcal M}(x,y,z)$ with respect to $x$, $y$, and $z$, where $d_2$ denotes the second order absolute difference using the mid point model, i.e. let $\mathcal C = \{c | \exists g(\cdot;x,z) : c = g(\frac{1}{2};x,z)\}$ the set of mid points between $x$ and $z$ on the manifold $\mathcal M$. Then the absolute second order difference is defined as

$ d2(x,y,z) = \min{c\in\mathcal C_{x,z}} d(c,y).$

While the (sub)gradient with respect to $y$ is easy, the other two require the evaluation of an adjointJacobiField. See Bačák, Bergmann, Steidl, Weinmann, 2016 for the derivation

source