JacobiFields

Jacobi Fields

A smooth tangent vector field $J\colon [0,1] \to T\mathcal M$ along a geodesic $g(\cdot;x,y)$ is called Jacobi field if it fulfills the ODE

\[\displaystyle 0 = \frac{D}{dt}J + R(J,\dot g)\dot g,\]

where $R$ is the Riemannian curvature tensor. Such Jacobi fields can be used to derive closed forms for the exponential map, the logarithmic map and the geodesic, all of them with respect to both arguments: Let $F\colon\mathcal N \to \mathcal M$ be given (for the $\exp_x\cdot$ we have $\mathcal N = T_x\mathcal M$, otherwise $\mathcal N=\mathcal M$) and denote by $\Xi_1,\ldots,\Xi_d$ an orthonormal frame along $g(\cdot;x,y)$ that diagonalizes the curvature tensor with corresponding eigenvalues $\kappa_1,\ldots,\kappa_d$. Note that on symmetric manifolds such a frame always exists.

Then $DF(x)[\eta] = \sum_{k=1}^d \langle \eta,\Xi_k(0)\rangle_x\beta(\kappa_k)\Xi_k(T)$ holds, where $T$ also depends on the function $F$ as the weights $\beta$. The values stem from solving the corresponding system of (decoupled) ODEs.

Note that in different references some factors might be a little different, for example when using unit speed geodesics.

The following weights functions are available

Manopt.βDgx — Function.

βDgx(κ,t,d)

weights for the jacobiField corresponding to the differential of the geodesic with respect to its start point $D_x g(t;x,y)[\eta]$. They are

\[\beta(\kappa) = \begin{cases} \frac{\sinh(d(1-t)\sqrt{-\kappa})}{\sinh(d\sqrt{-\kappa})} &\text{ if }\kappa < 0,\\ 1-t & \text{ if } \kappa = 0,\\ \frac{\sin((1-t)d\sqrt{\kappa})}{\sinh(d\sqrt{\kappa})} &\text{ if }\kappa > 0. \end{cases}\]

Due to a symmetry agrument, these are also used to compute $D_y g(t; x,y)[\eta]$

See also

DyGeo, DxGeo, jacobiField

source

Manopt.βDexpx — Function.

βDexpx(κ,t,d)

weights for the jacobiField corresponding to the differential of the geodesic with respect to its start point $D_x \exp_x(\xi)[\eta]$. They are

\[\beta(\kappa) = \begin{cases} \cosh(\sqrt{-\kappa})&\text{ if }\kappa < 0,\\ 1 & \text{ if } \kappa = 0,\\ \cos(\sqrt{\kappa}) &\text{ if }\kappa > 0. \end{cases}\]

See also

DxExp, jacobiField

source

Manopt.βDexpξ — Function.

βDexpξ(κ,t,d)

weights for the jacobiField corresponding to the differential of the geodesic with respect to its start point $D_\xi \exp_x(\xi)[\eta]$. They are

\[$\beta(\kappa) = \begin{cases} \frac{\sinh(d\sqrt{-\kappa})}{d\sqrt{-\kappa}}&\text{ if }\kappa < 0,\\ 1 & \text{ if } \kappa = 0,\\ \frac{\sin(d\sqrt{\kappa})}{\sqrt{d\kappa}}&\text{ if }\kappa > 0. \end{cases}\]

See also

DξExp, jacobiField

source

Manopt.βDlogx — Function.

βDlogx(κ,t,d)

weights for thejacobiField corresponding to the differential of the geodesic with respect to its start point $D_x \log_xy[\eta]$. They are

\[\beta(\kappa) = \begin{cases} -\sqrt{-\kappa}d\frac{\cosh(d\sqrt{-\kappa})}{\sinh(d\sqrt{-\kappa})}&\text{ if }\kappa < 0,\\ -1 & \text{ if } \kappa = 0,\\ -\sqrt{\kappa}d\frac{\cos(d\sqrt{\kappa})}{\sin(d\sqrt{\kappa})}&\text{ if }\kappa > 0. \end{cases}\]

See also

DxLog, DyLog, jacobiField

source

Manopt.βDlogy — Function.

βDlogy(κ,t,d)

weights for the JacobiField corresponding to the differential of the logarithmic map with respect to its argument $D_y \log_xy[\eta]$. They are

\[\beta(\kappa) = \begin{cases} \frac{ d\sqrt{-\kappa} }{\sinh(d\sqrt{-\kappa})}&\text{ if }\kappa < 0,\\ 1 & \text{ if } \kappa = 0,\\ \frac{ d\sqrt{\kappa} }{\sin(d\sqrt{\kappa})}&\text{ if }\kappa > 0. \end{cases}\]

See also

DyLog, jacobiField

source