Data

For some manifolds there are artificial or real application data available that can be loaded using the following data functions

Manopt.artificial_S1_signal — Function

artificial_S1_signal([pts=500])

generate a real-valued signal having piecewise constant, linear and quadratic intervals with jumps in between. If the resulting manifold the data lives on, is the Circle the data is also wrapped to $[-\pi,\pi)$.

Optional

pts – (500) number of points to sample the function

artificial_S1_signal(x)

evaluate the example signal $f(x), x ∈ [0,1]$, of phase-valued data introduces in Sec. 5.1 of

Bergmann, Laus, Steidl, Weinmann, Second Order Differences of Cyclic Data and Applications in Variational Denoising, SIAM J. Imaging Sci., 7(4), 2916–2953, 2014. doi: 10.1137/140969993

for values outside that intervall, this Signal is missing.

Manopt.artificial_S1_slope_signal — Function

artificial_S1_slope_signal([pts=500, slope=4.])

Creates a Signal of (phase-valued) data represented on the CircleManifold with increasing slope.

Optional

pts – (500) number of points to sample the function.

slope – (4.0) initial slope that gets increased afterwards

Manopt.artificialIn_SAR_image — Function

artificialIn_SAR_image([pts=500])

generate an artificial InSAR image, i.e. phase valued data, of size pts x pts points.

Manopt.artificial_S2_lemniscate — Function

artificial_S2_lemniscate(p [,pts=128,a=π/2,interval=[0,2π])

generate a Signal on the Sphere $\mathbb S^2$ by creating the Lemniscate of Bernoulli in the tangent space of p sampled at pts points and use exp to get a signal on the Sphere.

Input

p – the tangent space the Lemniscate is created in

pts – (128) number of points to sample the Lemniscate

a – (π/2) defines a half axis of the Lemniscate to cover a half sphere.

interval – ([0,2*π]) range to sample the lemniscate at, the default value refers to one closed curve

artificial_S2_lemniscate(p,t; a=π/2)

generate a point from the signal on the Sphere $\mathbb S^2$ by creating the Lemniscate of Bernoulli in the tangent space of p sampled at t and use èxp` to obtain a point on the Sphere.

Input

p – the tangent space the Lemniscate is created in

t – value to sample the Lemniscate at

Optional Values

a – (π/2) defines a half axis of the Lemniscate to cover a half sphere.

Manopt.artificial_S2_whirl_patch — Function

artificial_S2_whirl_patch([pts=5])

create a whirl within the pts$\times$pts patch of Sphere(@ref)(2)-valued image data.

Optional Parameters

pts – (5) size of the patch. If the number is odd, the center is the north pole.

Manopt.artificial_S2_whirl_image — Function

artificial_S2_whirl_image([pts=64])

generate an artificial image of data on the 2 sphere,

Arguments

pts – (64) size of the image in pts$\times$pts pixel.

Manopt.artificial_S2_rotation_image — Function

artificial_S2_rotation_image([pts=64, rotations=(.5,.5)])

creates an image with a rotation on each axis as a parametrization.

Optional Parameters

pts – (64) number of pixels along one dimension

rotations – ((.5,.5)) number of total rotations performed on the axes.

Manopt.artificial_SPD_image — Function

artificial_SPD_image([pts=64, stepsize=1.5])

create an artificial image of symmetric positive definite matrices of size pts$\times$pts pixel with a jump of size stepsize.

Manopt.artificial_SPD_image2 — Function

artificial_SPD_image2([pts=64, fraction=.66])

create an artificial image of symmetric positive definite matrices of size pts$\times$pts pixel with right hand side fraction is moved upwards.