Error Measures

Manopt.meanSquaredError — Function

meanSquaredError(M, p, q)

Compute the (mean) squared error between the two points p and q on the (power) manifold M.

Manopt.meanAverageError — Function

meanSquaredError(M,x,y)

Compute the (mean) squared error between the two points x and y on the PowerManifold manifold M.