Solvers

Solvers can be applied to Problems with solver specific Options.

List of Algorithms

The following algorithms are currently available

Solver	File	Problem & Option

Alternating Gradient Descent	alterating_gradient_descent.jl	AlternatingGradientProblem, AlternatingGradientDescentOptions

Chambolle-Pock	Chambolle-Pock.jl	PrimalDualProblem, ChambollePockOptions

Cyclic Proximal Point	cyclic_proximal_point.jl	ProximalProblem, CyclicProximalPointOptions

Douglas–Rachford	DouglasRachford.jl	ProximalProblem, DouglasRachfordOptions

Gradient Descent	gradient_descent.jl	GradientProblem, GradientDescentOptions

Nelder-Mead	NelderMead.jl	CostProblem, NelderMeadOptions

Particle Swarm	particle_swarm.jl	CostProblem, ParticleSwarmOptions

Quasi-Newton Method	quasi_newton.jl	GradientProblem, QuasiNewtonOptions

Subgradient Method	subgradient_method.jl	SubGradientProblem, SubGradientMethodOptions

Steihaug-Toint Truncated Conjugate-Gradient Method	truncated_conjugate_gradient_descent.jl	HessianProblem, TruncatedConjugateGradientOptions

The Riemannian Trust-Regions Solver	trust_regions.jl	HessianProblem, TrustRegionsOptions

Note that the solvers (or their Options to be precise) can also be decorated to enhance your algorithm by general additional properties, see debug output and recording values.

Technical Details

The main function a solver calls is

Manopt.solve — Method

solve(p,o)

run the solver implemented for the Problemp and the Optionso employing initialize_solver!, step_solver!, as well as the stop_solver! of the solver.

which is a framework, that you in general should not change or redefine. It uses the following methods, which also need to be implemented on your own algorithm, if you want to provide one.

Manopt.initialize_solver! — Function

initialize_solver!(p,o)

Initialize the solver to the optimization Problem by initializing all values in the Optionso.

Manopt.step_solver! — Function

step_solver!(p,o,iter)

Do one iteration step (the iterth) for Problemp by modifying the values in the Options o.

Manopt.get_solver_result — Function

get_solver_result(o)

Return the final result after all iterations that is stored within the (modified during the iterations) Options o.

Manopt.stop_solver! — Method

stop_solver!(p,o,i)

depending on the current Problem p, the current state of the solver stored in Options o and the current iterate i this function determines whether to stop the solver by calling the StoppingCriterion.