Frank—Wolfe method

Manopt.Frank_Wolfe_method — Function

Frank_Wolfe_method(M, f, grad_f, p)
Frank_Wolfe_method(M, gradient_objective, p; kwargs...)
Frank_Wolfe_method!(M, f, grad_f, p; kwargs...)
Frank_Wolfe_method!(M, gradient_objective, p; kwargs...)

Perform the Frank-Wolfe algorithm to compute for $\mathcal C \subset \mathcal M$ the constrained problem

\[ \operatorname*{arg\,min}_{p∈\mathcal C} f(p),\]

where the main step is a constrained optimisation is within the algorithm, that is the sub problem (Oracle)

\[ \operatorname*{arg\,min}_{q ∈ C} ⟨\operatorname{grad} f(p_k), \log_{p_k}q⟩.\]

for every iterate $p_k$ together with a stepsize $s_k≤1$. The algorhtm can be performed in-place of p.

This algorithm is inspired by but slightly more general than [WS22].

The next iterate is then given by $p_{k+1} = γ_{p_k,q_k}(s_k)$, where by default $γ$ is the shortest geodesic between the two points but can also be changed to use a retraction and its inverse.

Input

M a manifold $\mathcal M$

f: a cost function $f: \mathcal M→ℝ$ implemented as (M, p) -> v

grad_f: the gradient $\operatorname{grad}f: \mathcal M → T\mathcal M$ of f as a function (M, p) -> X or a function (M, X, p) -> X computing X in-place

p an initial value p $= p^{(0)} ∈ \mathcal M$

Alternatively to f and grad_f you can provide the AbstractManifoldGradientObjective gradient_objective directly.

Keyword arguments

evaluation=AllocatingEvaluation(): specify whether the functions that return an array, for example a point or a tangent vector, work by allocating its result (AllocatingEvaluation) or whether they modify their input argument to return the result therein (InplaceEvaluation). Since usually the first argument is the manifold, the modified argument is the second. For example grad_f(M,p) allocates, but grad_f!(M, X, p) computes the result in-place of X.

retraction_method=default_retraction_method(M, typeof(p)): a retraction $\operatorname{retr}$ to use, see the section on retractions.

stepsize=DecreasingStepsize(; length=2.0, shift=2): a functor inheriting from Stepsize to determine a step size, where the default is the step size $s_k = \frac{2}{k+2}$

stopping_criterion=StopAfterIteration(500)|StopWhenGradientNormLess(1.0e-6)) a functor inheriting from StoppingCriterion indicating when to stop.

X=zero_vector(M,p): specify a memory internally to store a tangent vector, the evaluated gradient $\operatorname{grad} f$ evaluated at $p^{(k)}$.

sub_cost=FrankWolfeCost(p, X): the cost of the Frank-Wolfe sub problem. This is used to define the sub_objective= keyword and has hence no effect, if you set sub_objective directly.

sub_grad=FrankWolfeGradient(p, X): the gradient of the Frank-Wolfe sub problem. This is used to define the sub_objective= keyword and has hence no effect, if you set sub_objective directly.

sub_kwargs=(;): a named tuple of keyword arguments that are passed to decorate_objective! of the sub solvers objective, the decorate_state! of the subsovlers state, and the sub state constructor itself.

sub_objective=ManifoldGradientObjective(sub_cost, sub_gradient): the objective for the Frank-Wolfe sub problem. This is used to define the sub_problem= keyword and has hence no effect, if you set sub_problem directly.

sub_problem=DefaultManoptProblem(M, sub_objective)): the sub problem to solve. This can be given in three forms

as an AbstractManoptProblem, then the sub_state= specifies the solver to use

as a closed form solution, as a function evaluating with new allocations (M, p, X) -> q that solves the sub problem on M given the current iterate p and (sub)gradient X.

as a closed form solution, as a function (M, q, p, X) -> q working in place of q.

For points 2 and 3 the sub_state has to be set to the corresponding AbstractEvaluationType, AllocatingEvaluation and InplaceEvaluation, respectively This keyword takes further into account sub_kwargs to evejtually decorate the problem

sub_state= if sub_problem isa Function evaluation else GradientDescentState(M, copy(M,p); kwargs...):

specify either the solver for a sub_problem or the kind of evaluation if the sub problem is given by a closed form solution this keyword takes into account the sub_stopping_criterion, and the sub_kwargs, that are also used to potentially decorate the state.

sub_stopping_criterion=StopAfterIteration(300)|StopWhenStepsizeLess(1e-8): a functor inheriting from StoppingCriterion indicating when to stop. for the sub solver. This is used to define the sub_state= keyword and has hence no effect, if you set sub_state directly.

All other keyword arguments are passed to decorate_state! for state decorators or decorate_objective! for objective, respectively.

If you provide the ManifoldGradientObjective directly, the evaluation= keyword is ignored. The decorations are still applied to the objective.

Output

the obtained (approximate) minimizer $p^*$, see get_solver_return for details

Manopt.Frank_Wolfe_method! — Function

Frank_Wolfe_method(M, f, grad_f, p)
Frank_Wolfe_method(M, gradient_objective, p; kwargs...)
Frank_Wolfe_method!(M, f, grad_f, p; kwargs...)
Frank_Wolfe_method!(M, gradient_objective, p; kwargs...)

Perform the Frank-Wolfe algorithm to compute for $\mathcal C \subset \mathcal M$ the constrained problem

\[ \operatorname*{arg\,min}_{p∈\mathcal C} f(p),\]

where the main step is a constrained optimisation is within the algorithm, that is the sub problem (Oracle)

\[ \operatorname*{arg\,min}_{q ∈ C} ⟨\operatorname{grad} f(p_k), \log_{p_k}q⟩.\]

for every iterate $p_k$ together with a stepsize $s_k≤1$. The algorhtm can be performed in-place of p.

This algorithm is inspired by but slightly more general than [WS22].

The next iterate is then given by $p_{k+1} = γ_{p_k,q_k}(s_k)$, where by default $γ$ is the shortest geodesic between the two points but can also be changed to use a retraction and its inverse.

Input

M a manifold $\mathcal M$

f: a cost function $f: \mathcal M→ℝ$ implemented as (M, p) -> v

grad_f: the gradient $\operatorname{grad}f: \mathcal M → T\mathcal M$ of f as a function (M, p) -> X or a function (M, X, p) -> X computing X in-place

p an initial value p $= p^{(0)} ∈ \mathcal M$

Alternatively to f and grad_f you can provide the AbstractManifoldGradientObjective gradient_objective directly.

Keyword arguments

evaluation=AllocatingEvaluation(): specify whether the functions that return an array, for example a point or a tangent vector, work by allocating its result (AllocatingEvaluation) or whether they modify their input argument to return the result therein (InplaceEvaluation). Since usually the first argument is the manifold, the modified argument is the second. For example grad_f(M,p) allocates, but grad_f!(M, X, p) computes the result in-place of X.

retraction_method=default_retraction_method(M, typeof(p)): a retraction $\operatorname{retr}$ to use, see the section on retractions.

stepsize=DecreasingStepsize(; length=2.0, shift=2): a functor inheriting from Stepsize to determine a step size, where the default is the step size $s_k = \frac{2}{k+2}$

stopping_criterion=StopAfterIteration(500)|StopWhenGradientNormLess(1.0e-6)) a functor inheriting from StoppingCriterion indicating when to stop.

X=zero_vector(M,p): specify a memory internally to store a tangent vector, the evaluated gradient $\operatorname{grad} f$ evaluated at $p^{(k)}$.

sub_cost=FrankWolfeCost(p, X): the cost of the Frank-Wolfe sub problem. This is used to define the sub_objective= keyword and has hence no effect, if you set sub_objective directly.

sub_grad=FrankWolfeGradient(p, X): the gradient of the Frank-Wolfe sub problem. This is used to define the sub_objective= keyword and has hence no effect, if you set sub_objective directly.

sub_kwargs=(;): a named tuple of keyword arguments that are passed to decorate_objective! of the sub solvers objective, the decorate_state! of the subsovlers state, and the sub state constructor itself.

sub_objective=ManifoldGradientObjective(sub_cost, sub_gradient): the objective for the Frank-Wolfe sub problem. This is used to define the sub_problem= keyword and has hence no effect, if you set sub_problem directly.

sub_problem=DefaultManoptProblem(M, sub_objective)): the sub problem to solve. This can be given in three forms

as an AbstractManoptProblem, then the sub_state= specifies the solver to use

as a closed form solution, as a function evaluating with new allocations (M, p, X) -> q that solves the sub problem on M given the current iterate p and (sub)gradient X.

as a closed form solution, as a function (M, q, p, X) -> q working in place of q.

For points 2 and 3 the sub_state has to be set to the corresponding AbstractEvaluationType, AllocatingEvaluation and InplaceEvaluation, respectively This keyword takes further into account sub_kwargs to evejtually decorate the problem

sub_state= if sub_problem isa Function evaluation else GradientDescentState(M, copy(M,p); kwargs...):

specify either the solver for a sub_problem or the kind of evaluation if the sub problem is given by a closed form solution this keyword takes into account the sub_stopping_criterion, and the sub_kwargs, that are also used to potentially decorate the state.

sub_stopping_criterion=StopAfterIteration(300)|StopWhenStepsizeLess(1e-8): a functor inheriting from StoppingCriterion indicating when to stop. for the sub solver. This is used to define the sub_state= keyword and has hence no effect, if you set sub_state directly.

All other keyword arguments are passed to decorate_state! for state decorators or decorate_objective! for objective, respectively.

If you provide the ManifoldGradientObjective directly, the evaluation= keyword is ignored. The decorations are still applied to the objective.

Output

the obtained (approximate) minimizer $p^*$, see get_solver_return for details

State

Manopt.FrankWolfeState — Type

FrankWolfeState <: AbstractManoptSolverState

A struct to store the current state of the Frank_Wolfe_method

It comes in two forms, depending on the realisation of the subproblem.

Fields

p: the current iterate $p=p^{(k)} ∈ \mathcal M$

X: the current gradient $\operatorname{grad}f(p^{(k)}) ∈ T_p\mathcal M$

inverse_retraction_method::AbstractInverseRetractionMethod : an inverse retraction $\operatorname{retr}^{-1}$

sub_problem::Union{AbstractManoptProblem, F}: a manopt problem or a function for a closed form solution of the sub problem

sub_state::Union{AbstractManoptSolverState,AbstractEvaluationType}: for a sub problem state which solver to use, for the closed form solution function, indicate, whether the closed form solution function works with AllocatingEvaluation) (M, p, X) -> q or with an InplaceEvaluation) (M, q, p, X) -> q

stop::StoppingCriterion : a functor indicating when to stop and whether the algorithm has stopped

stepsize::Stepsize : a stepsize.

retraction_method::AbstractRetractionMethod : a retraction $\operatorname{retr}: T\mathcal M \to \mathcal M$

The sub task requires a method to solve

\[ \operatorname*{arg\,min}_{q ∈ C} ⟨\operatorname{grad} f(p_k), \log_{p_k}q⟩.\]

Constructor

FrankWolfeState(M, p, sub_problem, sub_state; kwargs...)

Initialise the Frank Wolfe method state with.

Input

M a manifold $\mathcal M$

p an initial value p $= p^{(0)} ∈ \mathcal M$

X a tangent vector

sub_problem a AbstractManoptProblem to specify a problem for a solver or a closed form solution function.

sub_state a AbstractManoptSolverState for the sub_problem or a AbstractEvaluationType if a closed form solution is provided.

Keyword arguments

stopping_criterion=StopAfterIteration(200)|StopWhenGradientNormLess(1e-6) is used to set the field stop.

stepsize=default_stepsize(M, FrankWolfeState)

retraction_method=default_retraction_method(M, typeof(p))

inverse_retraction_method=default_inverse_retraction_method(M, typeof(p))

X=zero_vector(M,p)

where the remaining fields from before are keyword arguments.

Helpers

For the inner sub-problem you can easily create the corresponding cost and gradient using

Manopt.FrankWolfeCost — Type

FrankWolfeCost{P,T}

A structure to represent the oracle sub problem in the Frank_Wolfe_method. The cost function reads

\[F(q) = ⟨X, \log_p q⟩\]

The values p and X are stored within this functor and should be references to the iterate and gradient from within FrankWolfeState.

Manopt.FrankWolfeGradient — Type

FrankWolfeGradient{P,T}

A structure to represent the gradient of the oracle sub problem in the Frank_Wolfe_method, that is for a given point p and a tangent vector X the function reads

\[F(q) = ⟨X, \log_p q⟩\]

Its gradient can be computed easily using adjoint_differential_log_argument.

The values p and X are stored within this functor and should be references to the iterate and gradient from within FrankWolfeState.

[WS22]

M. Weber and S. Sra. Riemannian Optimization via Frank-Wolfe Methods. Mathematical Programming 199, 525–556 (2022).