Nelder–Mead

Nelder Mead Method

Manopt.NelderMead — Function.

NelderMead(M, F [, p])

perform a nelder mead minimization problem for the cost funciton F on the manifold M. If the initial population p is not given, a random set of points is chosen.

This algorithm is adapted from the Euclidean Nelder-Mead method, see https://en.wikipedia.org/wiki/Nelder–Mead_method and http://www.optimization-online.org/DB_FILE/2007/08/1742.pdf.

Input

M – a manifold $\mathcal M$

F – a cost function $F\colon\mathcal M\to\mathbb R$ to minimize

population – (n+1 randomMPoint(M)) an initial population of $n+1$ points, where $n$ is the dimension of the manifold M.

Optional

stoppingCriterion – (stopAfterIteration(2000)) a StoppingCriterion

retraction – (exp) a retraction(M,x,ξ) to use.

α – (1.) reflection parameter ($\alpha > 0$)

γ – (2.) expansion parameter ($\gamma$)

ρ – (1/2) contraction parameter, $0 < \rho \leq \frac{1}{2}$,

σ – (1/2) shrink coefficient, $0 < \sigma \leq 1$

and the ones that are passed to decorateOptions for decorators.

Output

either x the last iterate or the complete options depending on the optional keyword returnOptions, which is false by default (hence then only x is returned).

source

Options

Manopt.NelderMeadOptions — Type.

NelderMeadOptions <: Options

Describes all parameters and the state of a Nealer-Mead heuristic based optimization algorithm.

Fields

The naming of these parameters follows the Wikipedia article of the Euclidean case. The default is given in brackets, the required value range after the description

population – an Array{MPoint,1} of $n+1$ points $x_i$, $i=1,\ldots,n+1$, where $n$ is the dimension of the manifold.

stoppingCriterion – (stopAfterIteration(2000)) a StoppingCriterion

retraction – (exp) the rectraction to use

α – (1.) reflection parameter ($\alpha > 0$)

γ – (2.) expansion parameter ($\gamma>0$)

ρ – (1/2) contraction parameter, $0 < \rho \leq \frac{1}{2}$,

σ – (1/2) shrink coefficient, $0 < \sigma \leq 1$

x – (p[1]) - a field to collect the current best value

Constructors

NelderMead(M,stop, retr; α=1. , γ=2., ρ=1/2, σ=1/2)

construct a Nelder-Mead Option with a set of dimension(M)+1 random points.

NelderMead(p, stop retr; α=1. , γ=2., ρ=1/2, σ=1/2)

construct a Nelder-Mead Option with a set p of MPoints

source