Covariance matrix adaptation evolutionary strategy

The CMA-ES algorithm has been implemented based on [Han23] paper with basic Riemannian adaptations, related to transport of covariance matrix and its update vectors. Other attempts at adapting CMA-ES to Riemannian optimzation include [CFFS10].

Covariance matrix transport is handled by spd_matrix_transport_to function which is based on the idea of transport of matrix eigenvectors.

Manopt.cma_es — Function

cma_es(M, f, p_m=rand(M); σ::Real=1.0, kwargs...)

Perform covariance matrix adaptation evolutionary strategy search for global gradient-free randomized optimization. It is suitable for complicated non-convex functions. It can be reasonably expected to find global minimum within 3σ distance from p_m.

Implementation is based on [Han23] with basic adaptations to the Riemannian setting.

Input

M: a manifold $\mathcal M$

f: a cost function $f: \mathcal M→ℝ$ to find a minimizer $p^*$ for

Optional

p_m: (rand(M)) an initial point p

σ (1.0) initial standard deviation

λ (4 + Int(floor(3 * log(manifold_dimension(M))))population size (can be increased for a more thorough search)

tol_fun (1e-12) tolerance for the StopWhenPopulationCostConcentrated, similar to absolute difference between function values at subsequent points

tol_x (1e-12) tolerance for the StopWhenPopulationStuckConcentrated, similar to absolute difference between subsequent point but actually computed from distribution parameters.

Output

the obtained (approximate) minimizer $p^*$. To obtain the whole final state of the solver, see get_solver_return for details.

Manopt.spd_matrix_transport_to — Function

spd_matrix_transport_to(M::AbstractManifold, p, spd_coords, q, basis::AbstractBasis, vtm::AbstractVectorTransportMethod)

Transport the SPD matrix with spd_coords when expanded in basis from point p to point q on M.

(p, spd_coords) belongs to the fiber bundle of $B = \mathcal M × SPD(n)$, where n is the (real) dimension of M. The function corresponds to the Ehresmann connection defined by vector transport vtm of eigenvectors of spd_coords.

State

Manopt.CMAESState — Type

CMAESState{P,T} <: AbstractManoptSolverState

State of covariance matrix adaptation evolution strategy.

Stopping Criteria

Manopt.CMAESConditionCov — Type

CMAESConditionCov <: StoppingCriterion

Stop CMA-ES if condition number of covariance matrix exceeds threshold.

Manopt.StopWhenBestCostInGenerationConstant — Type

StopWhenBestCostInGenerationConstant <: StoppingCriterion

Stop if the range of the best objective function values of the last iteration_range generations is zero. This corresponds to EqualFUnValues condition from [Han23].

See also StopWhenPopulationCostConcentrated.

Manopt.StopWhenEvolutionStagnates — Type

StopWhenEvolutionStagnates{TParam<:Real} <: StoppingCriterion

The best and median fitness in each iteraion is tracked over the last 20% but at least min_size and no more than max_size iterations. Solver is stopped if in both histories the median of the most recent fraction of values is not better than the median of the oldest fraction.

Manopt.StopWhenPopulationCostConcentrated — Type

StopWhenPopulationCostConcentrated{TParam<:Real} <: StoppingCriterion

Stop if the range of the best objective function value in the last max_size generations and all function values in the current generation is below tol. This corresponds to TolFun condition from [Han23].

Constructor

StopWhenPopulationCostConcentrated(tol::Real, max_size::Int)

Manopt.StopWhenPopulationDiverges — Type

StopWhenPopulationDiverges{TParam<:Real} <: StoppingCriterion

Stop if σ times maximum deviation increased by more than tol. This usually indicates a far too small σ, or divergent behavior. This corresponds to TolXUp condition from [Han23].

Manopt.StopWhenPopulationStuckConcentrated — Type

StopWhenPopulationStuckConcentrated{TParam<:Real} <: StoppingCriterion

Stop if the standard deviation in all coordinates is smaller than tol and norm of σ * p_c is smaller than tol. This corresponds to TolX condition from [Han23].

Literature

[CFFS10]

S. Colutto, F. Fruhauf, M. Fuchs and O. Scherzer. The CMA-ES on Riemannian Manifolds to Reconstruct Shapes in 3-D Voxel Images. IEEE Transactions on Evolutionary Computation 14, 227–245 (2010).

[Han23]

N. Hansen. The CMA Evolution Strategy: A Tutorial (Mar 2023).