Conjugate Gradient Descent

Missing docstring.

Missing docstring for conjugate_gradient_descent. Check Documenter's build log for details.

Options

Manopt.ConjugateGradientDescentOptions — Type

ConjugateGradientOptions <: Options

specify options for a conjugate gradient descent algoritm, that solves a [GradientProblem].

Fields

x – the current iterate, a point on a manifold

∇ – the current gradient

δ – the current descent direction, i.e. also tangent vector

β – the current update coefficient rule, see .

coefficient – a DirectionUpdateRule function to determine the new β

stepsize – a Stepsize function

stop – a StoppingCriterion

retraction – a function to perform a step on the manifold

See also

conjugate_gradient_descent, GradientProblem, ArmijoLinesearch

Available Coefficients

Manopt.ConjugateDescentCoefficient — Type

ConjugateDescentCoefficient <: DirectionUpdateRule

Computes an update coefficient for the conjugate gradient method, where the ConjugateGradientDescentOptionso include the last iterates $x_k,\xi_k$, the current iterates $x_{k+1},\xi_{k+1}$ and the last update direction $\delta=\delta_k$, where the last three ones are stored in the variables with prequel Old based on

R. Fletcher, Practical Methods of Optimization vol. 1: Unconstrained Optimization John Wiley & Sons, New York, 1987. doi 10.1137/1024028

adapted to manifolds:

\[\beta_k = \frac{ \lVert \xi_{k+1} \rVert_{x_{k+1}}^2 } {\langle -\delta_k,\xi_k \rangle_{x_k}}.\]

See also: conjugate_gradient_descent

Manopt.DaiYuanCoefficient — Type

DaiYuanCoefficient <: DirectionUpdateRule

Computes an update coefficient for the conjugate gradient method, where the ConjugateGradientDescentOptionso include the last iterates $x_k,\xi_k$, the current iterates $x_{k+1},\xi_{k+1}$ and the last update direction $\delta=\delta_k$, where the last three ones are stored in the variables with prequel Old based on

[Y. H. Dai and Y. Yuan, A nonlinear conjugate gradient method with a strong global convergence property, SIAM J. Optim., 10 (1999), pp. 177–182. doi: 10.1137/S1052623497318992

adapted to manifolds: let $\nu_k = \xi_{k+1} - P_{x_k\to x_{k+1}}\xi_k$, where $P_{a\gets b}(\cdot)$ denotes a vector transport from the tangent space at $a$ to $b$.

Then the coefficient reads

\[\beta_k = \frac{ \lVert \xi_{k+1} \rVert_{x_{k+1}}^2 } {\langle P_{x_k\to x_{k+1}}\delta_k, \nu_k \rangle_{x_{k+1}}}.\]

See also: conjugate_gradient_descent

Constructor

DaiYuanCoefficient(t::AbstractVectorTransportMethod=ParallelTransport(), a::StoreOptionsAction=())

Construct the Dai Yuan coefficient update rule, where the parallel transport is the default vector transport and a new storage is created by default.

Manopt.DirectionUpdateRule — Type

DirectionUpdateRule

A general functor, that handles direction update rules. It's field(s) is usually only a StoreOptionsAction by default initialized to the fields required for the specific coefficient, but can also be replaced by a (common, global) individual one that provides these values.

Manopt.FletcherReevesCoefficient — Type

FletcherReevesCoefficient <: DirectionUpdateRule

Computes an update coefficient for the conjugate gradient method, where the ConjugateGradientDescentOptionso include the last iterates $x_k,\xi_k$, the current iterates $x_{k+1},\xi_{k+1}$ and the last update direction $\delta=\delta_k$, where the last three ones are stored in the variables with prequel Old based on

R. Fletcher and C. Reeves, Function minimization by conjugate gradients, Comput. J., 7 (1964), pp. 149–154. doi: 10.1093/comjnl/7.2.149

adapted to manifolds:

$ \betak = \frac{\lVert \xi{k+1}\rVert{x{k+1}}^2}{\lVert \xi{k}\rVert{x_{k}}^2}.$

See also: conjugate_gradient_descent

Manopt.HagerZhangCoefficient — Type

HagerZhangCoefficient <: DirectionUpdateRule

Computes an update coefficient for the conjugate gradient method, where the ConjugateGradientDescentOptionso include the last iterates $x_k,\xi_k$, the current iterates $x_{k+1},\xi_{k+1}$ and the last update direction $\delta=\delta_k$, where the last three ones are stored in the variables with prequel Old based on

[W. W. Hager and H. Zhang, A new conjugate gradient method with guaranteed descent and an efficient line search, SIAM J. Optim, (16), pp. 170-192, 2005. doi: 10.1137/030601880

adapted to manifolds: let $\nu_k = \xi_{k+1} - P_{x_k\to x_{k+1}}\xi_k$, where $P_{a\gets b}(\cdot)$ denotes a vector transport from the tangent space at $a$ to $b$.

\[\beta_k = \Bigl\langle\nu_k - \frac{ 2\lVert \nu_k\rVert_{x_{k+1}}^2 }{ \langle P_{x_k\to x_{k+1}}\delta_k, \nu_k \rangle_{x_{k+1}} } P_{x_k\to x_{k+1}}\delta_k, \frac{\xi_{k+1}}{ \langle P_{x_k\to x_{k+1}}\delta_k, \nu_k \rangle_{x_{k+1}} } \Bigr\rangle_{x_{k+1}}.\]

This method includes a numerical stability proposed by those authors.

See also: conjugate_gradient_descent

Constructor

HagerZhangCoefficient(t::AbstractVectorTransportMethod=ParallelTransport(), a::StoreOptionsAction=())

Construct the Hager Zhang coefficient update rule, where the parallel transport is the default vector transport and a new storage is created by default.

Manopt.HeestenesStiefelCoefficient — Type

HeestenesStiefelCoefficient <: DirectionUpdateRule

Computes an update coefficient for the conjugate gradient method, where the ConjugateGradientDescentOptionso include the last iterates $x_k,\xi_k$, the current iterates $x_{k+1},\xi_{k+1}$ and the last update direction $\delta=\delta_k$, where the last three ones are stored in the variables with prequel Old based on

M.R. Hestenes, E.L. Stiefel, Methods of conjugate gradients for solving linear systems, J. Research Nat. Bur. Standards, 49 (1952), pp. 409–436. doi: 10.6028/jres.049.044

adapted to manifolds as follows: let $\nu_k = \xi_{k+1} - P_{x_k\to x_{k+1}}\xi_k$. Then the update reads

\[\beta_k = \frac{\langle \xi_{k+1}, \nu_k \rangle_{x_{k+1}} } { \langle P_{x_{k+1}\gets x_k} \delta_k, \nu_k\rangle_{x_{k+1}} },\]

where $P_{a\gets b}(\cdot)$ denotes a vector transport from the tangent space at $a$ to $b$.

Constructor

HeestenesStiefelCoefficient(t::AbstractVectorTransportMethod=ParallelTransport(), a::StoreOptionsAction=())

Construct the Heestens Stiefel coefficient update rule, where the parallel transport is the default vector transport and a new storage is created by default.

See also: conjugate_gradient_descent

Manopt.LiuStoreyCoefficient — Type

LiuStoreyCoefficient <: DirectionUpdateRule

Computes an update coefficient for the conjugate gradient method, where the ConjugateGradientDescentOptionso include the last iterates $x_k,\xi_k$, the current iterates $x_{k+1},\xi_{k+1}$ and the last update direction $\delta=\delta_k$, where the last three ones are stored in the variables with prequel Old based on

[Y. Liu and C. Storey, Efficient generalized conjugate gradient algorithms, Part 1: Theory J. Optim. Theory Appl., 69 (1991), pp. 129–137. doi: 10.1007/BF00940464

adapted to manifolds: let $\nu_k = \xi_{k+1} - P_{x_{k+1}\gets x_k}\xi_k$, where $P_{a\gets b}(\cdot)$ denotes a vector transport from the tangent space at $a$ to $b$.

Then the coefficient reads

\[\beta_k = \frac{ \langle \xi_{k+1},\nu_k \rangle_{x_{k+1}} } {\langle -\delta_k,\xi_k \rangle_{x_k}}.\]

See also: conjugate_gradient_descent

Constructor

LiuStoreyCoefficient(t::AbstractVectorTransportMethod=ParallelTransport(), a::StoreOptionsAction=())

Construct the Lui Storey coefficient update rule, where the parallel transport is the default vector transport and a new storage is created by default.

Manopt.PolyakCoefficient — Type

PolyakCoefficient <: DirectionUpdateRule

Computes an update coefficient for the conjugate gradient method, where the ConjugateGradientDescentOptionso include the last iterates $x_k,\xi_k$, the current iterates $x_{k+1},\xi_{k+1}$ and the last update direction $\delta=\delta_k$, where the last three ones are stored in the variables with prequel Old based on

B. T. Polyak, The conjugate gradient method in extreme problems USSR Comp. Math. Math. Phys., 9 (1969), pp. 94–112. doi: 10.1016/0041-5553(69)90035-4

adapted to manifolds: let $\nu_k = \xi_{k+1} - P_{x_{k+1}\gets x_k}\xi_k$, where $P_{a\gets b}(\cdot)$ denotes a vector transport from the tangent space at $a$ to $b$.

Then the update reads

\[\beta_k = \frac{ \langle \xi_{k+1}, \nu_k \rangle_{x_{k+1}} } {\lVert \xi_k \rVert_{x_k} }.\]

Constructor

PolyakCoefficient(t::AbstractVectorTransportMethod=ParallelTransport(), a::StoreOptionsAction=())

Construct the Polyak coefficient update rule, where the parallel transport is the default vector transport and a new storage is created by default.

See also: conjugate_gradient_descent

Manopt.SteepestDirectionUpdateRule — Type

steepestDirectionUpdateRule <: DirectionUpdateRule

The simplest rule to update is to have no influence of the last direction and hence return an update $\beta = 0$ for all ConjugateGradientDescentOptionso

See also: conjugate_gradient_descent