Gradient descent

Manopt.gradient_descent — Function

gradient_descent(M, f, grad_f, p=rand(M); kwargs...)
gradient_descent(M, gradient_objective, p=rand(M); kwargs...)
gradient_descent!(M, f, grad_f, p; kwargs...)
gradient_descent!(M, gradient_objective, p; kwargs...)

perform the gradient descent algorithm

\[p_{k+1} = \operatorname{retr}_{p_k}\bigl( s_k\operatorname{grad}f(p_k) \bigr), \qquad k=0,1,…\]

where $s_k > 0$ denotes a step size.

The algorithm can be performed in-place of p.

Input

M a manifold $\mathcal M$

f: a cost function $f: \mathcal M→ℝ$ implemented as (M, p) -> v

grad_f: the gradient $\operatorname{grad}f: \mathcal M → T\mathcal M$ of f as a function (M, p) -> X or a function (M, X, p) -> X computing X in-place

p an initial value p $= p^{(0)} ∈ \mathcal M$

Alternatively to f and grad_f you can provide the AbstractManifoldGradientObjective gradient_objective directly.

Keyword arguments

direction=IdentityUpdateRule(): specify to perform a certain processing of the direction, for example Nesterov, MomentumGradient or AverageGradient.

evaluation=AllocatingEvaluation(): specify whether the functions that return an array, for example a point or a tangent vector, work by allocating its result (AllocatingEvaluation) or whether they modify their input argument to return the result therein (InplaceEvaluation). Since usually the first argument is the manifold, the modified argument is the second. For example grad_f(M,p) allocates, but grad_f!(M, X, p) computes the result in-place of X.

retraction_method=default_retraction_method(M, typeof(p)): a retraction $\operatorname{retr}$ to use, see the section on retractions.

stepsize=default_stepsize(M, GradientDescentState): a functor inheriting from Stepsize to determine a step size

stopping_criterion=StopAfterIteration(200)|StopWhenGradientNormLess(1e-8): a functor inheriting from StoppingCriterion indicating when to stop.

X=zero_vector(M,p): specify a memory internally to store a tangent vector, the evaluated gradient $\operatorname{grad} f$ evaluated at $p^{(k)}$.

All other keyword arguments are passed to decorate_state! for state decorators or decorate_objective! for objective, respectively.

If you provide the ManifoldGradientObjective directly, the evaluation= keyword is ignored. The decorations are still applied to the objective.

If you activate tutorial mode (cf. is_tutorial_mode), this solver provides additional debug warnings.

Output

The obtained approximate minimizer $p^*$. To obtain the whole final state of the solver, see get_solver_return for details, especially the return_state= keyword.

Manopt.gradient_descent! — Function

gradient_descent(M, f, grad_f, p=rand(M); kwargs...)
gradient_descent(M, gradient_objective, p=rand(M); kwargs...)
gradient_descent!(M, f, grad_f, p; kwargs...)
gradient_descent!(M, gradient_objective, p; kwargs...)

perform the gradient descent algorithm

\[p_{k+1} = \operatorname{retr}_{p_k}\bigl( s_k\operatorname{grad}f(p_k) \bigr), \qquad k=0,1,…\]

where $s_k > 0$ denotes a step size.

The algorithm can be performed in-place of p.

Input

M a manifold $\mathcal M$

f: a cost function $f: \mathcal M→ℝ$ implemented as (M, p) -> v

grad_f: the gradient $\operatorname{grad}f: \mathcal M → T\mathcal M$ of f as a function (M, p) -> X or a function (M, X, p) -> X computing X in-place

p an initial value p $= p^{(0)} ∈ \mathcal M$

Alternatively to f and grad_f you can provide the AbstractManifoldGradientObjective gradient_objective directly.

Keyword arguments

direction=IdentityUpdateRule(): specify to perform a certain processing of the direction, for example Nesterov, MomentumGradient or AverageGradient.

evaluation=AllocatingEvaluation(): specify whether the functions that return an array, for example a point or a tangent vector, work by allocating its result (AllocatingEvaluation) or whether they modify their input argument to return the result therein (InplaceEvaluation). Since usually the first argument is the manifold, the modified argument is the second. For example grad_f(M,p) allocates, but grad_f!(M, X, p) computes the result in-place of X.

retraction_method=default_retraction_method(M, typeof(p)): a retraction $\operatorname{retr}$ to use, see the section on retractions.

stepsize=default_stepsize(M, GradientDescentState): a functor inheriting from Stepsize to determine a step size

stopping_criterion=StopAfterIteration(200)|StopWhenGradientNormLess(1e-8): a functor inheriting from StoppingCriterion indicating when to stop.

X=zero_vector(M,p): specify a memory internally to store a tangent vector, the evaluated gradient $\operatorname{grad} f$ evaluated at $p^{(k)}$.

All other keyword arguments are passed to decorate_state! for state decorators or decorate_objective! for objective, respectively.

If you provide the ManifoldGradientObjective directly, the evaluation= keyword is ignored. The decorations are still applied to the objective.

If you activate tutorial mode (cf. is_tutorial_mode), this solver provides additional debug warnings.

Output

The obtained approximate minimizer $p^*$. To obtain the whole final state of the solver, see get_solver_return for details, especially the return_state= keyword.

State

Manopt.GradientDescentState — Type

GradientDescentState{P,T} <: AbstractGradientSolverState

Describes the state of a gradient based descent algorithm.

Fields

p: the current iterate $p=p^{(k)} ∈ \mathcal M$

X: the current gradient $\operatorname{grad}f(p^{(k)}) ∈ T_p\mathcal M$

stop::StoppingCriterion : a functor indicating when to stop and whether the algorithm has stopped

stepsize::Stepsize : a stepsize.

direction::DirectionUpdateRule : a processor to handle the obtained gradient and compute a direction to “walk into”.

retraction_method::AbstractRetractionMethod : a retraction $\operatorname{retr}: T\mathcal M \to \mathcal M$

Constructor

GradientDescentState(M, p=rand(M); kwargs...)

Initialize the gradient descent solver state, where

Input

M a manifold $\mathcal M$

p an initial value p $= p^{(0)} ∈ \mathcal M$

Keyword arguments

direction=IdentityUpdateRule()

stopping_criterion=StopAfterIteration(100) is used to set the field stop.

stepsize=default_stepsize(M, GradientDescentState; retraction_method=retraction_method)

retraction_method=default_retraction_method(M, typeof(p))

X=zero_vector(M,p)

See also

gradient_descent

Direction update rules

A field of the options is the direction, a DirectionUpdateRule, which by default IdentityUpdateRule just evaluates the gradient but can be enhanced for example to

Manopt.DirectionUpdateRule — Type

DirectionUpdateRule

A general functor, that handles direction update rules. It's fields are usually only a StoreStateAction by default initialized to the fields required for the specific coefficient, but can also be replaced by a (common, global) individual one that provides these values.

Manopt.IdentityUpdateRule — Type

IdentityUpdateRule <: DirectionUpdateRule

The default gradient direction update is the identity, usually it just evaluates the gradient.

Manopt.MomentumGradient — Type

MomentumGradient <: DirectionUpdateRule

Append a momentum to a gradient processor, where the last direction and last iterate are stored and the new is composed as $η_i = m*η_{i-1}' - s d_i$, where $sd_i$ is the current (inner) direction and $η_{i-1}'$ is the vector transported last direction multiplied by momentum $m$.

Fields

p_old: remember the last iterate for parallel transporting the last direction

momentum: factor for momentum

direction: internal DirectionUpdateRule to determine directions to add the momentum to.

vector_transport_method: vector transport method to use

X_old: the last gradient/direction update added as momentum

Constructors

Add momentum to a gradient problem, where by default just a gradient evaluation is used

MomentumGradient(
    M::AbstractManifold;
    p=rand(M),
    s::DirectionUpdateRule=IdentityUpdateRule();
    X=zero_vector(M, p),
    momentum=0.2
    vector_transport_method=default_vector_transport_method(M, typeof(p)),
)

Initialize a momentum gradient rule to s, where p and X are memory for interim values.

Manopt.AverageGradient — Type

AverageGradient <: DirectionUpdateRule

Add an average of gradients to a gradient processor. A set of previous directions (from the inner processor) and the last iterate are stored, average is taken after vector transporting them to the current iterates tangent space.

Fields

gradients: the last n gradient/direction updates

last_iterate: last iterate (needed to transport the gradients)

direction: internal DirectionUpdateRule to determine directions to apply the averaging to

vector_transport_method: vector transport method to use

Constructors

AverageGradient(
    M::AbstractManifold,
    p::P=rand(M);
    n::Int=10
    s::DirectionUpdateRule=IdentityUpdateRule();
    gradients = fill(zero_vector(p.M, o.x),n),
    last_iterate = deepcopy(x0),
    vector_transport_method = default_vector_transport_method(M, typeof(p))
)

Add average to a gradient problem, where

n: determines the size of averaging

s: is the internal DirectionUpdateRule to determine the gradients to store

gradients: can be pre-filled with some history

last_iterate: stores the last iterate

vector_transport_method: determines how to transport all gradients to the current iterates tangent space before averaging

Manopt.Nesterov — Type

Nesterov <: DirectionUpdateRule

Fields

γ

μ the strong convexity coefficient

v (=$=v_k$, $v_0=x_0$) an interim point to compute the next gradient evaluation point y_k

shrinkage (= i -> 0.8) a function to compute the shrinkage $β_k$ per iterate.

Assume $f$ is $L$-Lipschitz and $μ$-strongly convex. Given

a step size $h_k<\frac{1}{L}$ (from the GradientDescentState

a shrinkage parameter $β_k$

and a current iterate $x_k$

as well as the interim values $γ_k$ and $v_k$ from the previous iterate.

This compute a Nesterov type update using the following steps, see [ZS18]

Compute the positive root $α_k∈(0,1)$ of $α^2 = h_k\bigl((1-α_k)γ_k+α_k μ\bigr)$.

Set $\bar γ_k+1 = (1-α_k)γ_k + α_kμ$

$y_k = \operatorname{retr}_{x_k}\Bigl(\frac{α_kγ_k}{γ_k + α_kμ}\operatorname{retr}^{-1}_{x_k}v_k \Bigr)$

$x_{k+1} = \operatorname{retr}_{y_k}(-h_k \operatorname{grad}f(y_k))$

$v_{k+1} = `\operatorname{retr}_{y_k}\Bigl(\frac{(1-α_k)γ_k}{\barγ_k}\operatorname{retr}_{y_k}^{-1}(v_k) - \frac{α_k}{\bar γ_{k+1}}\operatorname{grad}f(y_k) \Bigr)$

$γ_{k+1} = \frac{1}{1+β_k}\bar γ_{k+1}$

Then the direction from $x_k$ to $x_k+1$ by $d = \operatorname{retr}^{-1}_{x_k}x_{k+1}$ is returned.

Constructor

Nesterov(M::AbstractManifold, p::P; γ=0.001, μ=0.9, shrinkage = k -> 0.8;
    inverse_retraction_method=LogarithmicInverseRetraction())

Initialize the Nesterov acceleration, where x0 initializes v.

Debug actions

Manopt.DebugGradient — Type

DebugGradient <: DebugAction

debug for the gradient evaluated at the current iterate

Constructors

DebugGradient(; long=false, prefix= , format= "$prefix%s", io=stdout)

display the short (false) or long (true) default text for the gradient, or set the prefix manually. Alternatively the complete format can be set.

Manopt.DebugGradientNorm — Type

DebugGradientNorm <: DebugAction

debug for gradient evaluated at the current iterate.

Constructors

DebugGradientNorm([long=false,p=print])

display the short (false) or long (true) default text for the gradient norm.

DebugGradientNorm(prefix[, p=print])

display the a prefix in front of the gradient norm.

Manopt.DebugStepsize — Type

DebugStepsize <: DebugAction

debug for the current step size.

Constructors

DebugStepsize(;long=false,prefix="step size:", format="$prefix%s", io=stdout)

display the a prefix in front of the step size.

Record actions

Manopt.RecordGradient — Type

RecordGradient <: RecordAction

record the gradient evaluated at the current iterate

Constructors

RecordGradient(ξ)

initialize the RecordAction to the corresponding type of the tangent vector.

Manopt.RecordGradientNorm — Type

RecordGradientNorm <: RecordAction

record the norm of the current gradient

Manopt.RecordStepsize — Type

RecordStepsize <: RecordAction

record the step size

Technical details

The gradient_descent solver requires the following functions of a manifold to be available

A retract!(M, q, p, X); it is recommended to set the default_retraction_method to a favourite retraction. If this default is set, a retraction_method= does not have to be specified.

By default gradient descent uses ArmijoLinesearch which requires max_stepsize(M) to be set and an implementation of inner(M, p, X).

By default the stopping criterion uses the norm as well, to stop when the norm of the gradient is small, but if you implemented inner, the norm is provided already.

By default the tangent vector storing the gradient is initialized calling zero_vector(M,p).

Literature

[Lue72]

D. G. Luenberger. The gradient projection method along geodesics. Management Science 18, 620–631 (1972).

[ZS18]

H. Zhang and S. Sra. Towards Riemannian accelerated gradient methods, arXiv Preprint, 1806.02812 (2018).